### 2023-03-30 #5314 @editorialbot
[PRE REVIEW]: lpjmlkit: A toolkit for operating LPJmL and model-specific data processing
**Submitting author:** @jnnsbrr (<a href="http://orcid.org/0000-0002-9055-6904">Jannes Breier</a>)
**Repository:** https://github.com/PIK-LPJmL/lpjmlkit
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v.1.0.13
