### 2023-03-30 #5314 @kthyng
Hi @jnnsbrr and thanks for your submission. Can you add a sentence or two to the beginning of the paper to explain more about dynamic global vegetative models and where they fit in with other models? There are a lot of acronyms and not a lot of explanation for context. Also is it normal to have to install R packages to see any docs?
