+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: lpjmlkit
 Type: Package
 Title: Toolkit for Basic LPJmL Handling
-Version: 1.0.13
+Version: 1.1.0
 Authors@R: c(
     person("Jannes", "Breier", , "jannesbr@pik-potsdam.de", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-9055-6904")),
     person("Sebastian","Ostberg", , "ostberg@pik-potsdam.de", role = "aut", comment = c(ORCID = "0000-0002-2368-7015")),
@@ -39,7 +39,8 @@ Imports:
     methods,
     abind,
     rlang,
-    withr
+    withr,
+    grDevices
 Suggests:
     rmarkdown,
     knitr,
@@ -50,4 +51,4 @@ Suggests:
     maps
 Config/testthat/edition: 3
 VignetteBuilder: knitr
-Date: 2023-03-27
+Date: 2023-03-29
