### 2023-03-31 f199fe6
add citation for dgvm definition
