+++ paper.bib
@@ -265,3 +265,21 @@ Publisher: Nature Publishing Group},
 	date = {2018-04-12},
 	note = {Publisher: Copernicus {GmbH}}
 }
+
+@article{cramer_global_2001,
+	title = {Global response of terrestrial ecosystem structure and function to {CO}2 and climate change: results from six dynamic global vegetation models},
+	volume = {7},
+	issn = {1365-2486},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1046/j.1365-2486.2001.00383.x},
+	doi = {10.1046/j.1365-2486.2001.00383.x},
+	shorttitle = {Global response of terrestrial ecosystem structure and function to {CO}2 and climate change},
+	abstract = {The possible responses of ecosystem processes to rising atmospheric {CO}2 concentration and climate change are illustrated using six dynamic global vegetation models that explicitly represent the interactions of ecosystem carbon and water exchanges with vegetation dynamics. The models are driven by the {IPCC} {IS}92a scenario of rising {CO}2 (Wigley et al. 1991), and by climate changes resulting from effective {CO}2 concentrations corresponding to {IS}92a, simulated by the coupled ocean atmosphere model {HadCM}2-{SUL}. Simulations with changing {CO}2 alone show a widely distributed terrestrial carbon sink of 1.4–3.8 Pg C y−1 during the 1990s, rising to 3.7–8.6 Pg C y−1 a century later. Simulations including climate change show a reduced sink both today (0.6–3.0 Pg C y−1) and a century later (0.3–6.6 Pg C y−1) as a result of the impacts of climate change on {NEP} of tropical and southern hemisphere ecosystems. In all models, the rate of increase of {NEP} begins to level off around 2030 as a consequence of the ‘diminishing return’ of physiological {CO}2 effects at high {CO}2 concentrations. Four out of the six models show a further, climate-induced decline in {NEP} resulting from increased heterotrophic respiration and declining tropical {NPP} after 2050. Changes in vegetation structure influence the magnitude and spatial pattern of the carbon sink and, in combination with changing climate, also freshwater availability (runoff). It is shown that these changes, once set in motion, would continue to evolve for at least a century even if atmospheric {CO}2 concentration and climate could be instantaneously stabilized. The results should be considered illustrative in the sense that the choice of {CO}2 concentration scenario was arbitrary and only one climate model scenario was used. However, the results serve to indicate a range of possible biospheric responses to {CO}2 and climate change. They reveal major uncertainties about the response of {NEP} to climate change resulting, primarily, from differences in the way that modelled global {NPP} responds to a changing climate. The simulations illustrate, however, that the magnitude of possible biospheric influences on the carbon balance requires that this factor is taken into account for future scenarios of atmospheric {CO}2 and climate change.},
+	pages = {357--373},
+	number = {4},
+	journaltitle = {Global Change Biology},
+	author = {Cramer, Wolfgang and Bondeau, Alberte and Woodward, F. Ian and Prentice, I. Colin and Betts, Richard A. and Brovkin, Victor and Cox, Peter M. and Fisher, Veronica and Foley, Jonathan A. and Friend, Andrew D. and Kucharik, Chris and Lomas, Mark R. and Ramankutty, Navin and Sitch, Stephen and Smith, Benjamin and White, Andrew and Young-Molling, Christine},
+	urldate = {2023-03-31},
+	date = {2001},
+	langid = {english},
+	keywords = {dynamic global vegetation model, global carbon cycle},
+}
