### 2023-03-31 #5314 @jnnsbrr
Dear @kthyng,
thank you! I added a definition for DGVMs to the Statement of need section (update is on the way).
Yes, its normal to first install R packages to be able to view the rendered rd files (in `./man`) in R.
For our R package we additionally provide the full documentation at [CRAN](https://cran.r-project.org/web/packages/lpjmlkit/lpjmlkit.pdf) and at [r-universe](https://pik-piam.r-universe.dev/lpjmlkit). Both are linked in the paper, CRAN publication in the [Summary](https://github.com/PIK-LPJmL/lpjmlkit/blob/JOSS/paper.md#:~:text=lpjmlkit%20R%20package-,%5B%40lpjmlkit_manual%5D,-is%20an%20open) and r-universe documentation in the [Documentation & License section](https://github.com/PIK-LPJmL/lpjmlkit/blob/JOSS/paper.md#:~:text=https%3A//pik%2Dpiam.r%2Duniverse.dev/lpjmlkit).
