+++ R/LPJmLData_as.R
@@ -134,12 +134,13 @@ LPJmLData$set("private",
 #' }
 #'
 #' @md
-#' @export
-as_tibble <- function(x,
-                      subset = NULL,
-                      aggregate = NULL,
-                      value_name = "value",
-                      ...) {
+#' @exportS3Method tibble::as_tibble
+as_tibble.LPJmLData <- function(x, # nolint
+                                subset = NULL,
+                                aggregate = NULL,
+                                value_name = "value",
+                                ...) {
+
   y <- x$as_tibble(subset,
                    aggregate,
                    value_name,
@@ -169,8 +170,9 @@ LPJmLData$set("private",
 #' Coerce an LPJmLData object to a Raster* object
 #'
 #' Function to coerce (convert) an [`LPJmLData`] object into a
-#' \link[raster]{raster} or \link[raster]{brick} object that allows for any
-#' GIS-based raster operations. Read more about the raster package at
+#' [RasterLayer](\link[raster]{raster}) or [RasterBrick](\link[raster]{brick})
+#' object that allows for any GIS-based raster operations.
+#' Read more about the raster package at
 #' <https://rspatial.github.io/raster/reference/raster-package.html>.
 #' The successor package of raster is called terra: <https://rspatial.org/>.
 #'
@@ -311,8 +313,8 @@ LPJmLData$set("private",
 #' Coerce an LPJmLData object to a terra object
 #'
 #' Function to coerce (convert) an [`LPJmLData`] object into a
-#' \link[terra]{rast} object that allows GIS-based raster operations.
-#' Read more about the terra package at <https://rspatial.org/>.
+#' [SpatRaster](\link[terra]{rast}) object that allows GIS-based raster
+#' operations. Read more about the terra package at <https://rspatial.org/>.
 #'
 #' @param x [LPJmLData] object.
 #'
@@ -351,6 +353,7 @@ LPJmLData$set("private",
 #' }
 #'
 #' @md
+#' @aliases as_rast as_SpatRaster
 #' @export
 as_terra <- function(x,
                      subset = NULL,
