+++ man/as_tibble.LPJmLData.Rd
@@ -1,10 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/LPJmLData_as.R
-\name{as_tibble}
-\alias{as_tibble}
+\name{as_tibble.LPJmLData}
+\alias{as_tibble.LPJmLData}
 \title{Coerce an LPJmLData object to a tibble}
 \usage{
-as_tibble(x, subset = NULL, aggregate = NULL, value_name = "value", ...)
+as_tibble.LPJmLData(
+  x,
+  subset = NULL,
+  aggregate = NULL,
+  value_name = "value",
+  ...
+)
 }
 \arguments{
 \item{x}{\link{LPJmLData} object}
