+++ NAMESPACE
@@ -6,13 +6,13 @@ S3method(length,LPJmLData)
 S3method(plot,LPJmLData)
 S3method(subset,LPJmLData)
 S3method(summary,LPJmLData)
+S3method(tibble::as_tibble,LPJmLData)
 export(add_grid)
 export(as_array)
 export(as_header)
 export(as_list)
 export(as_raster)
 export(as_terra)
-export(as_tibble)
 export(asub)
 export(calc_cellarea)
 export(check_config)
