+++ NAMESPACE
@@ -1,12 +1,12 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(as_tibble,LPJmLData)
 S3method(dim,LPJmLData)
 S3method(dimnames,LPJmLData)
 S3method(length,LPJmLData)
 S3method(plot,LPJmLData)
 S3method(subset,LPJmLData)
 S3method(summary,LPJmLData)
-S3method(tibble::as_tibble,LPJmLData)
 export(add_grid)
 export(as_array)
 export(as_header)
@@ -34,3 +34,4 @@ export(write_config)
 export(write_header)
 importFrom(foreach,"%dopar%")
 importFrom(magrittr,"%>%")
+importFrom(tibble,as_tibble)
