### 2023-04-12 #5314 @kthyng
Hi @jnnsbrr — What sort of expertise would overlap with the topic of your software? We seem to be having some trouble finding a good fit.
