### 2023-04-17 f2cc580
Fix cran, adjust runner vignette & more flexible read_config
