+++ paper.md
@@ -308,9 +308,10 @@ where users can also report issues and suggest improvements.
 We would like to thank Pascal Führlich and Jan Philipp Dietrich for providing
 access to PIK Research Software Engineering resources and additional support
 during the development process.
-Special thanks also go to Werner von Bloh for implementing new features into
+Special thanks go to Werner von Bloh for implementing new features into
 LPJmL to enable the development of lpjmlkit, such as writing meta
 files.
+We would also like to thank the Earth4All project for the support that made this development possible.  
 Further thanks go to Susanne Rolinski and Marie Hemmen who supported the project
 along the way.
 
