+++ R/LPJmLData_as.R
@@ -171,8 +171,8 @@ LPJmLData$set("private",
 #' Coerce an LPJmLData object to a raster object
 #'
 #' Function to coerce (convert) an [`LPJmLData`] object into a
-#' [RasterLayer](\link[raster]{raster}) or [RasterBrick](\link[raster]{brick})
-#' object that allows for any GIS-based raster operations.
+#' \link[raster]{raster} or \link[raster]{brick} object that allows for any
+#' GIS-based raster operations.
 #' Read more about the raster package at
 #' <https://rspatial.github.io/raster/reference/raster-package.html>.
 #' The successor package of raster is called terra: <https://rspatial.org/>.
@@ -314,7 +314,7 @@ LPJmLData$set("private",
 #' Coerce an LPJmLData object to a terra object
 #'
 #' Function to coerce (convert) an [`LPJmLData`] object into a
-#' [SpatRaster](\link[terra]{rast}) object that allows GIS-based raster
+#' \link[terra]{rast} object that allows GIS-based raster
 #' operations. Read more about the terra package at <https://rspatial.org/>.
 #'
 #' @param x [LPJmLData] object.
