### 2023-04-19 #5314 @kthyng
@jnnsbrr We have a pool of editors for JOSS from which we need to choose, but for reviewers we can pick essentially anyone who can/will do the job and doesn't have a conflict of interest! Your reviewer suggestions will be very helpful for @graciellehigino in choosing reviewers.
