### 2023-05-22 #5447 @tonyewong
I had a read through the accompanying paper, which is well-written and helpful. I know JOSS papers are supposed to not include too much of the API descriptions, which instead should be in the software documentation on GitHub. I'm not sure but I'm curious whether some areas of the paper might be better left to more extensive documentation in GitHub? For example lines 78-89 or 93-101.
