### 2023-05-22 #5447 @tonyewong
Also related to the paper -- are there other models that can be cited, beyond MAgPIE and LPJ-GUESS, that have a similar focus? Are there other similarities/differences that can be noted in the paper?
