### 2023-05-31 #5447 @jnnsbrr
Thanks a lot @graciellehigino and @tonyewong!

And yes, @tonyewong, you are right - it should not include too much of the API descriptions as mentioned [here](https://joss.readthedocs.io/en/latest/submitting.html#:~:text=Given%20this%20format%2C%20a%20%E2%80%9Cfull%20length%E2%80%9D%20paper%20is%20not%20permitted%2C%20and%20software%20documentation%20such%20as%20API%20(Application%20Programming%20Interface)%20functionality%20should%20not%20be%20in%20the%20paper%20and%20instead%20should%20be%20outlined%20in%20the%20software%20documentation). I rather wanted to highlight the key features that define *lpjmlkit*, so one get a better understanding of what is an LPJmL kit and what is meant by that in R.
I will try to abstract that more and rather give example usages for ongoing research projects instead.

Regarding other models that can be cited - MAgPIE and LPJ-GUESS are both good examples, one from a different discipline, but same research institute and one from the same discipline but another research institute.Both examples have a similar story of how research has been conducted over the last decade and what tools have had to be developed to keep it state of the art.  Therefore, I do not believe that adding more examples will contribute to a better understanding of this issue.

Can you also already tick the boxes that can easily be excluded from the scope of this work, such as data sharing and research on humans and animals? 🙂
Please let me know if there is anything else I can do to move this forward.

Many thanks in advance 🙏
