### 2023-06-28 #5447 @tonyewong
Hello all, and apologies for the (big!) delay! We had summer interns starting over the past couple weeks so things were chaotic. But I'm back, and alive, as far as I can tell...!

I'm working on installing and running some tests. Installing goes smoothly (great!) and I see where in the GitHub repository there is a `tests` directory with relevant `output` directories too. Are these what the vignettes are referring to in their relative file paths? Can some further instructions be added about how to run some example cases and benchmark tests?

It looks like everything we need is already in the GitHub repo, so this is probably pretty straightforward to do without the user needing to clone the GitHub.
