### 2023-07-05 3cc7d14
adjust tests for joss reviewer and update readme for cran on lpjml installation
