+++ DESCRIPTION
@@ -51,4 +51,4 @@ Suggests:
     maps
 Config/testthat/edition: 3
 VignetteBuilder: knitr
-Date: 2023-07-05
+Date: 2023-07-06
