### 2023-07-06 de2f842
adjust tests for joss reviewer and update readme for cran on lpjml installation
