### 2023-07-07 b9f4413
rewrite package features part as requested
