+++ paper.bib
@@ -283,3 +283,19 @@ Publisher: Nature Publishing Group},
 	langid = {english},
 	keywords = {dynamic global vegetation model, global carbon cycle},
 }
+
+@article{wickham_tidy_2014,
+	title = {Tidy Data},
+	volume = {59},
+	rights = {Copyright (c) 2013 Hadley  Wickham},
+	issn = {1548-7660},
+	url = {https://doi.org/10.18637/jss.v059.i10},
+	doi = {10.18637/jss.v059.i10},
+	abstract = {A huge amount of effort is spent cleaning data to get it ready for analysis, but there has been little research on how to make data cleaning as easy and effective as possible. This paper tackles a small, but important, component of data cleaning: data tidying. Tidy datasets are easy to manipulate, model and visualize, and have a specific structure: each variable is a column, each observation is a row, and each type of observational unit is a table. This framework makes it easy to tidy messy datasets because only a small set of tools are needed to deal with a wide range of un-tidy datasets. This structure also makes it easier to develop tidy tools for data analysis, tools that both input and output tidy datasets. The advantages of a consistent data structure and matching tools are demonstrated with a case study free from mundane data manipulation chores.},
+	pages = {1--23},
+	journaltitle = {Journal of Statistical Software},
+	author = {Wickham, Hadley},
+	urldate = {2023-07-07},
+	date = {2014-09-12},
+	langid = {english},
+}
