+++ paper.md
@@ -154,116 +154,65 @@ run LPJmL model simulations as well as read and process the resulting data.
 ## LPJmL Runner
 
 The Runner module is designed to operate LPJmL on Unix-based operating systems
-that have a working LPJmL installation and includes four key functions.
+that have a working LPJmL version 4 or higher installation and includes four
+key functions.
 
-- `write_config()` generates JSON configuration files by utilizing a
-base configuration file as a source template to be preprocessed prior to use,
-along with a data frame containing parameters to be modified. It also sets up
-the directory structure and files to execute the runs and collect the outputs.
+The principle is based on the procedure of creating tables for the parameters
+and settings to be changed for a large number of related simulations,
+e.g. multi-scenario or uncertainty/sensitivity analyses.
+The creation of the table is based on the concept of tidy data with columns as
+parameters/settings (variables) and rows as simulations (observations)
+[@wickham_tidy_2014].
+In order to facitlitate the execution of these simulations, `write_config`
+handles the transformation of corresponding tables into LPJmL compatible JSON
+configuration files.
 
-- `check_config()` tests whether generated config.json files are valid
-for LPJmL simulations using LPJmL utility functionality.
-
-- `run_lpjml()` executes LPJmL as a subprocess directly from within
-the current R session. This is particularly useful for simulations with a small
-number of spatial cells that require few resources.
-
-- `submit_lpjml()` function submits the LPJmL Simulations to SLURM to be run
-by an HPC cluster using additional resources and detached from the current
-R session, which is necessary for (global) runs that include all or many spatial
-cells to be simulated.
-
-The following is a typical example of the use of *LPJmL Runner*:
-```R
-library("lpjmlkit")
-
-lpjml_path <- "./LPJmL_internal"
-sim_path <- "./simulations"
-
-# Define data frame with configuration parameters to be changed
-config_params <- data.frame(
-  sim_name = c("spinup", "landuse", "natural_vegetation"),
-  landuse = c("no", "yes", "no"),
-  reservoir = c(FALSE, TRUE, FALSE),
-  river_routing = c(FALSE, TRUE, FALSE),
-  wateruse = c("no", "yes", "no"),
-  const_deposition = c(FALSE, FALSE, TRUE),
-  dependency = c(NA, "spinup", "spinup")
-)
-
-# Write corresponding configuration files using a (precompiled) base
-# configuration file
-config_details <- write_config(config_params,
-                               model_path = lpjml_path,
-                               sim_path = sim_path)
-
-# Check validity of each written configuration file
-check_config(config_details,
-             model_path = lpjml_path,
-             sim_path = sim_path)
-
-# Submit LPJmL simulations to SLURM on an HPC cluster
-submit_lpjml(config_details,
-             model_path = lpjml_path,
-             sim_path = sim_path)
-```
+LPJmL simulations can either be run sequentially locally with `run_lpjml()`
+which is particularly suitable for testing the model or for focusing on small
+regions.
+In addition, simulations can be submitted to an HPC cluster equipped with the
+SLURM workload manager using `submit_lpjml()`, which is particularly useful for
+computing multiple global simulations in parallel.
 
 ## LPJmL Data
 
-The Data module provides various functions for reading and processing LPJmL
-data.
+While LPJmL Runner covers the whole range of running LPJmL simulations, LPJmL
+Data provides the tools for the subsequent data analysis part.
+LPJmL simulations use their own binary data format to write memory efficient
+output data.
+The associated metadata can be written either as a file header or as an
+additional metadata file in JSON format.
+Using `read_io()`, this data can be read into a standardised data format called
[…]
