### 2023-07-07 #5447 @jnnsbrr
Dear @tonyewong,

thank you! I rewrote the **Package features** part as promised earlier.

The `tests` directory ensures all functions work as intended via unit tests with coverage above 90%. As lpjmlkit is just an interface for the LPJmL model, I did not include any data-driven benchmarks.
As global data analysis requires large files to be exchanged this would be rather difficult, if you intended performance benchmarks?
And the vignette for LPJmLData is not written for the data in the `tests/output` folder but can be used for it (with different results). It instead provides exemplary paths that would be created by simulations using the LPJmL Runner.

I hope that clarifies some points.
Thanks a lot!
