### 2023-07-07 #5447 @tonyewong
Thanks for clarifying!

Can the documentation be clarified to let the reader know exactly how to run some benchmark tests to verify that the software is running as expected? For example, comparing to some standard benchmark output? This does not necessarily need to be global in scale, if it would be easier to do for a reduced spatial domain.

I'm thinking specifically about these three items in the checklist:

1. **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
2. **Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
3. **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Thanks again!
