### 2023-07-10 075887a
fix encoding warnings, fix test failures
