### 2023-07-10 #5447 @jnnsbrr
Dear @tonyewong,

thanks! I will try to explain it in more detail:

1. [lpjmlkit](https://github.com/PIK-LPJmL/lpjmlkit) follows the standard way of providing example usages via the [vignettes](https://github.com/PIK-LPJmL/lpjmlkit#tutorial), see them also automatically rendered in the [r-universe repository](https://pik-piam.r-universe.dev/lpjmlkit) [for LPJmL Data](https://pik-piam.r-universe.dev/articles/lpjmlkit/lpjml-data.html) and [for LPJmL Runner](https://pik-piam.r-universe.dev/articles/lpjmlkit/lpjml-runner.html) - the base for these files are the [Rmd files on github](https://github.com/PIK-LPJmL/lpjmlkit/tree/master/vignettes) - they show common application examples.
2. The whole documentation is [also automatically rendered at r-universe](https://pik-piam.r-universe.dev/lpjmlkit/doc/manual.html#lpjmlkit-package) - still the base for that is the roxygen documentation within each exported function of the package on github also following the standard procedure for R packages
3. Automated tests are provided with the unit tests of lpjmlkit, stored in the `./tests` folder. You can run them for example via `devtools::test("<path_to_lpjmlkit>")` also following common practices. The Codecov tool provides a good overview how much of the functions is covered  by tests: https://app.codecov.io/gh/PIK-LPJmL/lpjmlkit/tree/master/R

Further (script-based) benchmarks are not part of the standard procedure when developing and publishing R packages, see also  https://r-pkgs.org/testing-basics.html.
Still you could apply the examples described in the vignettes (mainly [LPJmLData](https://pik-piam.r-universe.dev/articles/lpjmlkit/lpjml-data.html) to the data provided in the `.tests/testdata/output` folder (other outputs), even though these are examples for simulations with only three 0.5° cells.
Or I would have to provide additional data stored in a cloud service so you can test it on a global dataset to get a better visual impression - but I have not seen something like that as a request for a JOSS article (?).

I hope that helps!

Thanks a lot!
