### 2023-07-10 #5447 @tonyewong
Sure, I understand that and that makes sense. What I'm thinking about is including guidance to readers to [objectively assess whether the software works](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) once they've installed the package. I don't see anything explicit in the software right now. But what you said as #3 there I think would satisfy the requirements I linked above, with just a mention of that in the README!
