### 2023-07-10 #5447 @jnnsbrr
Ah - good point - thank you.  Citing JOSS ...
> Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)

See https://github.com/PIK-LPJmL/lpjmlkit#readme > the check badge of the header: This should in theory be "passing", unfortunately the GitHub Actions pipeline (hosted at my Institute) currently has a problem with encodings, which is why it fails (its not the lpjmlkit package - and it did work before). We are currently working on a solution! If you click on the badge, you see that it was passing the check already before.
