### 2023-07-11 8bdca76
fix encoding problems with stringi, fix test failures with file.copy/remove...
