+++ R/utils.R
@@ -67,18 +67,26 @@ names_recursively <- function(x) {
 }
 
 
-# colorize variable name for messages, warning, stop
+# colorize variable name for messages, warning, stop in blue
 col_var <- function(x) {
   cli::col_blue(x)
 }
 
+# colorize notes in yellow
 col_note <- function(x) {
   cli::col_yellow(x)
 }
 
+# colorize warnings and important strings in red
 col_warn <- function(x) {
   cli::col_red(x)
 }
+
+# bold strings, especially for headers of messages
+bold_head <- function(x) {
+  cli::style_bold(x)
+}
+
 
 
 
