### 2023-07-13 9e0a0f9
following Stephens suggestions, and changing slurm_options to slurm_option to not use plural forms for arguments
