+++ R/write_config.R
@@ -121,7 +121,7 @@
 #' in the SLURM queue. This can be achieved by supplying this option as a
 #' parameter to `param`. \cr
 #' 6 options are available, namely `sclass`, `ntask`, `wtime`, `blocking`,
-#' `constraint` and `slurm_options`. Use as arguments for [submit_lpjml()`].\cr
+#' `constraint` and `slurm_option`. Use as arguments for [submit_lpjml()`].\cr
 #' If specified in `param`, they overwrite the corresponding function arguments
 #' in [`submit_lpjml()`].
 #'
@@ -315,7 +315,7 @@ write_config <- function(x,
                                dependency = NA)
 
   slurm_args <- c(
-    "sclass", "ntask", "wtime", "blocking", "constraint", "slurm_options"
+    "sclass", "ntask", "wtime", "blocking", "constraint", "slurm_option"
   )
 
   config_tmp[slurm_args] <- NA
