+++ paper.md
@@ -142,70 +142,71 @@ calibration, benchmarking or indicator development.
 # Package features
 
 *lpjmlkit*  is an R package that contains two main modules, each of which
-contains supplementary documentation in the form of a guide in addition to the
+contains supplementary documentation in form of a guide in addition to the
 package documentation:
 
 * [*LPJmL Runner* vignette](https://pik-piam.r-universe.dev/articles/lpjmlkit/lpjml-runner.html)
 * [*LPJmL Data* vignette](https://pik-piam.r-universe.dev/articles/lpjmlkit/lpjml-data.html)
 
-Together, these modules can be used to set up and
+Together, these modules can be used to configure and
 run LPJmL model simulations as well as read and process the resulting data.
 
 ## LPJmL Runner
 
 The Runner module is designed to operate LPJmL on Unix-based operating systems
-that have a working LPJmL version 4 or higher installation and includes four
+that have a working LPJmL installation (version 4 or higher) and includes four
 key functions.
 
-The principle is based on the procedure of creating tables for the parameters
-and settings to be changed for a large number of related simulations,
+The basic idea is to create tables for the parameters
+and settings to be changed from a default configuration file
+for single runs or any number of related simulations,
 e.g. multi-scenario or uncertainty/sensitivity analyses.
-The creation of the table is based on the concept of tidy data with columns as
+The creation of these tables is based on the concept of `tidy data` with columns as
 parameters/settings (variables) and rows as simulations (observations)
 [@wickham_tidy_2014].
-In order to facitlitate the execution of these simulations, `write_config`
-handles the transformation of corresponding tables into LPJmL compatible JSON
+In order to facitlitate the execution of these simulations, the function `write_config()`
+handles the formatting of the corresponding `tidy data` tables into LPJmL compatible JSON
 configuration files.
 
-LPJmL simulations can either be run sequentially locally with `run_lpjml()`
+LPJmL simulations can either be run sequentially on single CPUs with `run_lpjml()`,
 which is particularly suitable for testing the model or for focusing on small
 regions.
 In addition, simulations can be submitted to an HPC cluster equipped with the
 SLURM workload manager using `submit_lpjml()`, which is particularly useful for
-computing multiple global simulations in parallel.
+computing multiple global simulations in parallel mode using multiple CPUs.
 
 ## LPJmL Data
 
-While LPJmL Runner covers the whole range of running LPJmL simulations, LPJmL
-Data provides the tools for the subsequent data analysis part.
-LPJmL simulations use their own binary data format to write memory efficient
-output data.
-The associated metadata can be written either as a file header or as an
+While the module *LPJmL Runner* covers the whole range of running LPJmL simulations, 
+`LPJmL Data` provides the tools for the subsequent data analysis part.
+LPJmL simulations output data in a raw binary data format designed for memory efficient
+data structures.
+The LPJmL model can write metadata for each output file, either as a file header or as an
 additional metadata file in JSON format.
-Using `read_io()`, this data can be read into a standardised data format called
-`LPJmLData`, which associates the data arrays with the corresponding metadata
-to ensure its integrity.
+Using `read_io()`, LPJmL raw output data can be read into a standardised data format called
+`LPJmLData`, which makes use of the corresponding metadata
+to correctly represent the data arrays (dimensions, units, etc.).
 
-The core of this data format is the range of functionalities commonly used by
+This data format is designed to faciliitate the functionalities commonly used by
 LPJmL users for data analysis.
-In addition to the descriptive statistics that an LPJmLData object displays,
-there is a `plot()` method to get direct graphical insights into the data.
-LPJmLData objects can be transformed into various common spatial and temporal
-formats (`transform()`).
[…]
