### 2023-07-13 6413c9f
adjust doc for submit_lpjml, ntask -> ntasks also in write_config
