### 2023-07-18 04d55fa
new options for submitting lpjml to SLURM
