+++ R/submit_lpjml.R
@@ -27,17 +27,32 @@
 #'   More information at <https://www.pik-potsdam.de/en>. Defaults to `"short"`.
 #'
 #' @param ntasks Integer defining the number of tasks/threads. More information
-#'   at <https://www.pik-potsdam.de/en> and <https://slurm.schedmd.com>.
-#'   Defaults to `256`.
+#'   at <https://www.pik-potsdam.de/en> and
+#'   <https://slurm.schedmd.com/sbatch.html>. Defaults to `256`.
 #'
 #' @param wtime Character string defining the time limit. Setting a lower time
 #'   limit than the maximum runtime for `sclass` can reduce the wait time in the
 #'   SLURM job queue. More information at <https://www.pik-potsdam.de/en> and
-#'   <https://slurm.schedmd.com>.
+#'   <https://slurm.schedmd.com/sbatch.html>.
 #'
 #' @param blocking Integer defining the number of cores to be blocked. More
 #'   information at <https://www.pik-potsdam.de/en> and
-#'   <https://slurm.schedmd.com>.
+#'   <https://slurm.schedmd.com/sbatch.html>.
+#'
+#' @param constraint Character string defining constraints for node selection.
+#'   Use `constraint = "haswell"` to request nodes of the type haswell with 16
+#'   cores per node, `constraint = "broadwell"` to request nodes of the type
+#'   broadwell CPUs with 32 cores per node or `constraint = "exclusive"` to
+#'   reserve all CPUs of assigned nodes even if less are requested by `ntasks`.
+#'   Using `exclusive` should prevent interference of other batch jobs with
+#'   LPJmL. More information at <https://www.pik-potsdam.de> and
+#'   <https://slurm.schedmd.com/sbatch.html>.
+
+#'
+#' @param slurm_options A named list of further arguments to be passed to sbatch.
+#'   E.g. list(`mail-user` = "max.mustermann@pik-potsdam.de")
+#'   More information at <https://www.pik-potsdam.de> and
+#'   <https://slurm.schedmd.com/sbatch.html>
 #'
 #' @param no_submit Logical. Set to `TRUE` to test if `x` set correctly or
 #'   `FALSE` to actually submit job to SLURM.
@@ -67,11 +82,11 @@
 #' | scen2_transient | scen1 _spinup  |
 #'
 #' To use different SLURM settings for each run the optional SLURM options
-#' `"sclass"`, `"ntask"`, `"wtime"` or `"blocking"` can also be supplied to the
-#' initial \link[tibble]{tibble} supplied as `param` to
+#' `"sclass"`, `"ntasks"`, `"wtime"`, "blocking"` or `constraint` can also be
+#' supplied to the initial \link[tibble]{tibble} supplied as `param` to
 #' [`write_config()`]. These overwrite the (default) SLURM
-#' arguments (`sclass`, `ntask`, `wtime` or `blocking`) supplied to
-#' `submit_lpjml`.
+#' arguments (`sclass`, `ntasks`, `wtime`, `blocking` or ` `constraint`)
+#' supplied to `submit_lpjml`.
 #'
 #' | **sim_name**    | **dependency** | **wtime** |
 #' |:--------------- |:-------------- |----------:|
@@ -175,6 +190,8 @@ submit_lpjml <- function(x, # nolint:cyclocomp_linter.
                          ntasks = 256,
                          wtime = "",
                          blocking = "",
+                         constraint = "",
+                         slurm_options = list(),
                          no_submit = FALSE,
                          output_path = NULL) {
 
@@ -213,7 +230,9 @@ submit_lpjml <- function(x, # nolint:cyclocomp_linter.
   x$type <- "simulation"
   x$job_id <- NA
[…]
