### 2023-08-25 #5447 @tonyewong
Hello! Apologies again for the delay, but I think I'm just about converged and done with my review.

The tests, functionality, documentation, etc all look like they're running great! Many thanks again for pointing out where folks can go to see the 0.5-degree small data sets they can use if one wishes to interactively run some tests or play with the code. The vignettes are all great and run smoothly.

The last comment I have is regarding this one:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Specifically, #2 and 3 are covered nicely, but there is not a note about #1 yet, at least not that I can see in the top-level README.md file. Can such a note be added? After that I think I can check my last box.

Well done!
