### 2023-08-29 #5447 @tonyewong
Sure, I think that would work. Alternatively, since you have #2 and 3 covered in the "Questions / Problems" section, you could add a statement about contributions there too. Any of these that you find aesthetically pleasing is okay from my end!
