### 2023-09-13 #5447 @jnnsbrr
> Sure, I think that would work. Alternatively, since you have #2 and 3 covered in the "Questions / Problems" section, you could add a statement about contributions there too. Any of these that you find aesthetically pleasing is okay from my end!

Dear @tonyewong, I just remembered that we already have a sentence for contributions **Documentation & License** section:
> The source code is available on GitHub at
[https://github.com/PIK-LPJmL/lpjmlkit](https://github.com/PIK-LPJmL/lpjmlkit),
where users can also report issues and suggest improvements.

Maybe I did not fully understood if you mean the paper or the repository's README . Regarding the README This part of the README is auto generated by an R package software developer team at PIK and follows specific conventions (https://pik-piam.r-universe.dev). Also [other iconic R packages](https://github.com/tidyverse/dplyr) do not explicitly mention contributiors.
