### 2023-09-13 #5447 @tonyewong
Right! I saw that in the JOSS paper. What I'm thinking about is in the repository itself, that it would be useful to include guidelines about how to contribute and report problems. Since it's in the paper, maybe this is fine? I don't know whether to interpret JOSS's requirement of this information as requiring it in the repository itself, or just _somewhere_ in either the paper or repository. I'm happy to follow @graciellehigino 's advice on this last review box :)
