### 2023-09-13 #5447 @graciellehigino
Hey!

@jnnsbrr the best practice overall is to have a separate document in the repo with details about how to contribute (do we fork and PR? Do we just open an issue? Do we open an issue and link a PR to it? See an example [here](https://opensource.com/life/16/3/contributor-guidelines-template-and-tips))

If you can do that, that'd be awesome, but technically we can check that box because you say on the paper that people should refer to the repo and in the repo you have a contact info, which is _something_ and it's useful.

I hope that helps!
