+++ tests/testthat/test-LPJmLData.R
@@ -169,10 +169,6 @@ test_that("test print method", {
     print(output),
     "meta"
   )
-  expect_output(
-    print(output),
-    "subset"
-  )
 
   # Check if data printed
   expect_output(
