+++ R/LPJmLMetaData.R
@@ -75,7 +75,7 @@ LPJmLMetaData <- R6::R6Class( # nolint
       cat(
         paste0(spaces,
                col_var(
-                paste0("$", print_fields)
+                 paste0("$", print_fields)
                ),
                " ",
                to_char1,
@@ -106,16 +106,19 @@ LPJmLMetaData <- R6::R6Class( # nolint
       )
       cat("\n")
 
-      cat(
-        paste0(
-          spaces,
-          col_var("$subset"),
-          " ",
-          # Color red if subset.
-          ifelse(self$subset, col_warn(self$subset), ""),
-          "\n"
+      # Print information about subset if subsetted
+      if (self$subset) {
+        cat(
+          paste0(
+            spaces,
+            col_var("$subset"),
+            " ",
+            # Color red if subset.
+            col_warn(self$subset),
+            "\n"
+          )
         )
-      )
+      }
     },
 
 
@@ -170,7 +173,7 @@ LPJmLMetaData <- R6::R6Class( # nolint
       # Update cell fields - distinguish between character -> LPJmL C index
       #   starting from 0 and numeric/integer -> R index starting from 1 -> -1.
       if (!is.null(subset$cell) ||
-          !is.null(subset$lon) || !is.null(subset$lat)) {
+            !is.null(subset$lon) || !is.null(subset$lat)) {
 
         # Subset of subset$cell, subset$lon or subset$lat always have to be
         #   accompanied by cell_dimnames.
@@ -276,39 +279,39 @@ LPJmLMetaData <- R6::R6Class( # nolint
             "bigendian" = ifelse(x$endian == "big", TRUE, FALSE),
             # "descr" = tolower(x$name), # nolint
             "lastyear" = x$header[["firstyear"]] +
-                         x$header[["timestep"]] *
-                         (x$header[["nyear"]] - 1),
+              x$header[["timestep"]] *
+                (x$header[["nyear"]] - 1),
             "name" = ifelse(is.null(x$name), "LPJDUMMY", x$name)
           )) %>%
           `[[<-`("order",
-                switch(as.character(.$order),
-                       `1` = "cellyear",
-                       `2` = "yearcell",
-                       `3` = "cellindex",
-                       `4` = "cellseq",
-                       stop(
-                         paste(
-                           "Invalid order value", sQuote(.$order), "in header"
-                         )
-                       )
-                  )
-                ) %>%
-          `[[<-`("datatype",
-                switch(as.character(.$datatype),
-                       `0` = "byte",
-                       `1` = "short",
-                       `2` = "int",
-                       `3` = "float",
-                       `4` = "double",
-                       stop(
-                         paste(
-                           "Invalid datatype value", sQuote(.$datatype),
-                           "in header"
-                         )
-                       )
-                    )
+            switch(as.character(.$order),
+              `1` = "cellyear",
+              `2` = "yearcell",
+              `3` = "cellindex",
+              `4` = "cellseq",
+              stop(
+                paste(
+                  "Invalid order value", sQuote(.$order), "in header"
                 )
-          private$init_list(header_to_meta, additional_attributes)
+              )
+            )
+          ) %>%
+          `[[<-`("datatype",
+            switch(as.character(.$datatype),
+              `0` = "byte",
+              `1` = "short",
+              `2` = "int",
+              `3` = "float",
+              `4` = "double",
+              stop(
+                paste(
+                  "Invalid datatype value", sQuote(.$datatype),
+                  "in header"
+                )
+              )
+            )
+          )
+        private$init_list(header_to_meta, additional_attributes)
 
       } else {
[…]
