### 2023-10-06 1853343
adjust behaviour, fix linter problems
