### 2023-10-19 e5779f2
small edits to CONTRIBUTING.md
