### 2023-11-08 8240f07
adjust for setting for file ending in LPJmL
