### 2023-11-09 #5447 @jnnsbrr
Dear @graciellehigino,
after some discussions, we added a [CONTRIBUTING.md](https://github.com/PIK-LPJmL/lpjmlkit/blob/JOSS/CONTRIBUTING.md) to the repository to help people who want to contribute to the project. I hope that matches the requirements.
Please let me know if there is anything else standing in the way of publication.

Thanks in advance 🙂
