### 2023-11-10 722eac2
enable add_grid method for variable named 'cellid'
