### 2023-11-10 c38bcce
buildlibrary and version number increment
