### 2023-11-16 f0237d2
speedup read io in two places
