+++ R/read_io.R
@@ -183,7 +183,9 @@ read_io <- function( # nolint:cyclocomp_linter.
   silent       = FALSE
 ) {
   # Switch off fancy quotes but revert setting when leaving the function
-  quotes_option <- options(useFancyQuotes = FALSE) # nolint:undesirable_function_linter.
+  # scipen = 999 is to ensure that cell or band indices are not written in
+  # scientific notation.
+  quotes_option <- options(useFancyQuotes = FALSE, scipen = 999) # nolint:undesirable_function_linter.
   on.exit(options(quotes_option)) # nolint:undesirable_function_linter.
 
   # Detect file_type if not provided by user
@@ -711,15 +713,11 @@ read_io_data <- function(
     )
 
     # Assign dimension names to array.
-    # Ensure cell or band indices are not written in scientific notation.
     band_names <- default(
       meta_data$band_names, seq_len(default(meta_data$nbands, 1))
-    ) %>%
-      format(trim = TRUE, scientific = FALSE, justify = "none")
+    )
 
-    options(scipen = 999) # effectively disable scientific notation for dimnames
     cell_dimnames <- seq(default(meta_data$firstcell, 0), length.out = meta_data$ncell)
-    options(scipen = 0)  # reset to default
 
     dimnames(year_data) <- switch(
       default(meta_data$order, "cellyear"),
