### 2023-11-16 1615467
subset check moved to subset function
