### 2023-11-23 bb0f395
Implement more useful error message when user tries to set locked attribute
