+++ R/LPJmLMetaData.R
@@ -327,111 +327,130 @@ LPJmLMetaData <- R6::R6Class( # nolint
   active = list(
 
     #' @field sim_name Simulation name (works as identifier in LPJmL Runner).
-    sim_name = function() {
+    sim_name = function(...) {
+      check_change(self, "sim_name", ...)
       return(private$.sim_name)
     },
 
     #' @field source LPJmL version (character string).
-    source = function() {
+    source = function(...) {
+      check_change(self, "source", ...)
       return(private$.source)
     },
 
     #' @field history Character string of the call used to run LPJmL. This
     #'   normally includes the path to the LPJmL executable and the path to the
     #'   configuration file for the simulation.
-    history = function() {
+    history = function(...) {
+      check_change(self, "history", ...)
       return(private$.history)
     },
 
     #' @field variable Name of the input/output variable, e.g. `"npp"` or
     #'   `"runoff"`.
-    variable = function() {
+    variable = function(...) {
+      check_change(self, "variable", ...)
       return(private$.variable)
     },
 
     #' @field descr Description of the input/output variable.
-    descr = function() {
+    descr = function(...) {
+      check_change(self, "descr", ...)
       return(private$.descr)
     },
 
     #' @field unit Unit of the input/output variable.
-    unit = function() {
+    unit = function(...) {
+      check_change(self, "unit", ...)
       return(private$.unit)
     },
 
     #' @field nbands Number (numeric) of bands (categoric dimension). Please
     #'   note that `nbands` follows the convention in LPJmL, which uses the
     #'   plural form for bands as opposed to `nyear` or `ncell`.
-    nbands = function() {
+    nbands = function(...) {
+      check_change(self, "nbands", ...)
       return(private$.nbands)
     },
 
     #' @field band_names Name of the bands (categoric dimension). Not included
     #'   if `nbands = 1`.
-    band_names = function() {
+    band_names = function(...) {
+      check_change(self, "band_names", ...)
       return(private$.band_names)
     },
 
     #' @field nyear Number (numeric) of data years in the parent `LPJmLData`
     #'   object.
-    nyear = function() {
+    nyear = function(...) {
+      check_change(self, "nyear", ...)
       return(private$.nyear)
     },
 
     #' @field firstyear First calendar year (numeric) in the parent `LPJmLData`
     #'   object.
-    firstyear = function() {
+    firstyear = function(...) {
+      check_change(self, "firstyear", ...)
       return(private$.firstyear)
     },
 
     #' @field lastyear Last calendar year (numeric) in the parent `LPJmLData`
     #'   object.
-    lastyear = function() {
+    lastyear = function(...) {
+      check_change(self, "lastyear", ...)
       return(private$.lastyear)
     },
 
     #' @field nstep Number (numeric) of intra-annual time steps. `1` for annual,
     #' `12` for monthly, and `365` for daily data.
-    nstep = function() {
+    nstep = function(...) {
+      check_change(self, "nstep", ...)
       return(private$.nstep)
     },
 
     #' @field timestep Number (numeric) of years between time steps.
     #'   `timestep = 5` means that output is written every 5 years.
-    timestep = function() {
+    timestep = function(...) {
+      check_change(self, "timestep", ...)
       return(private$.timestep)
     },
 
     #' @field ncell Number (numeric) of cells in the parent `LPJmLData` object.
-    ncell = function() {
+    ncell = function(...) {
+      check_change(self, "ncell", ...)
       return(private$.ncell)
     },
 
     #' @field firstcell First cell (numeric) in the parent `LPJmLData` object.
-    firstcell = function() {
+    firstcell = function(...) {
+      check_change(self, "firstcell", ...)
       return(private$.firstcell)
[…]
