### 2023-11-23 28719ce
Revert "buildlibrary and version number increment"
This reverts commit c38bccea3441ad633b9add143db21290289bdd89.
