+++ .github/workflows/check.yaml
@@ -23,7 +23,6 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: |
-            gamstransfer=?ignore
             any::lucode2
             any::covr
             any::madrat
@@ -57,8 +56,6 @@ jobs:
 
       - name: Test coverage
         shell: Rscript {0}
-        run: |
-          nonDummyTests <- setdiff(list.files("./tests/testthat/"), c("test-dummy.R", "_snaps"))
-          if(length(nonDummyTests) > 0) covr::codecov(quiet = FALSE)
+        run: covr::codecov(quiet = FALSE)
         env:
           NOT_CRAN: "true"
