### 2023-11-23 58fa751
Fix initialize function for LPJmLGridData to accept variable name cellid used in some new LPJmL versions instead of grid.
