### 2023-11-24 aa8bd1f
result after build library executed
