### 2023-12-05 fd6acfe
enable setting of file extension via netcdf branch
