### 2023-12-05 cbbe6fc
add support for new cjson files with new arg to write_config - cjson_filename. New error handling for non existing js/cjson config file.
