+++ R/write_config.R
@@ -17,7 +17,7 @@
 #'
 #' @param output_list Character vector containing the `"id"` of outputvars.
 #'   If defined, only these defined outputs will be written. Otherwise, all
-#'   outputs set in `js_filename` will be written. Defaults to `NULL`.
+#'   outputs set in `cjson_filename` will be written. Defaults to `NULL`.
 #'
 #' @param output_list_timestep Single character string or character vector
 #'   defining what temporal resolution the defined outputs from `output_list`
@@ -29,8 +29,8 @@
 #'   Defaults to `"raw"`. Further options: `"cdf"` (NetCDF) or `"clm"`
 #'   (file with header).
 #'
-#' @param js_filename Character string providing the name of the main js file to
-#'   be parsed. Defaults to `"lpjml.js"`.
+#' @param cjson_filename Character string providing the name of the main cjson
+#'   config file to be parsed. Defaults to `"lpjml_config.cjson"`.
 #'
 #' @param parallel_cores Integer defining the number of available CPU cores for
 #'   parallelization. Defaults to `4`.
@@ -44,6 +44,9 @@
 #' @param output_path Argument is deprecated as of version 1.0; use sim_path
 #'   instead.
 #'
+#' @param js_filename Argument is deprecated as of version 1.3; use
+#'   cjson_filename instead.
+#'
 #' @return \link[tibble]{tibble} with at least one column named `"sim_name"`.
 #'   Run parameters `"order"` and `"dependency"` are included if defined in
 #'   `x`. \link[tibble]{tibble} in this format is required for
@@ -53,9 +56,9 @@
 #'
 #' Supply a \link[tibble]{tibble} for `x`, in which each row represents
 #' a configuration (config) for an LPJmL simulation. \cr
-#' Here a config is referred to as the precompiled `"lpjml.js"` file (or file
-#' name provided as `js_filename` argument), which links to all other
-#' mandatory ".js" files. The precompilation is done internally by
+#' Here a config is referred to as the precompiled `"lpjml_config.cjson"` file
+#' (or file name provided as `cjson_filename` argument), which links to all
+#' other mandatory ".js" files. The precompilation is done internally by
 #' [`write_config()`].\cr
 #' `write_config()` uses the column names of `param` as keys for the config
 #' json using the same syntax as lists, e.g. `"k_temp"` from `"param.js"`
@@ -266,11 +269,12 @@ write_config <- function(x,
                          output_list = c(),
                          output_list_timestep = "annual",
                          output_format = "raw",
-                         js_filename = "lpjml.js",
+                         cjson_filename = "lpjml_config.json",
                          parallel_cores = 4,
                          debug = FALSE,
                          params = NULL,
-                         output_path = NULL) {
+                         output_path = NULL,
+                         js_filename = NULL) {
 
 
   # Deprecate argument params
@@ -288,6 +292,26 @@ write_config <- function(x,
   sim_path <- deprecate_arg(new_arg = sim_path,
                             deprec_arg = output_path,
                             version = "1.0.0")
+
+  # Deprecate argument js_filename
+  if (!is.null(js_filename)) {
+    cjson_filename <- deprecate_arg(
+      new_arg = cjson_filename,
+      deprec_arg = js_filename,
+      version = "1.3.0",
+      ignore_new_arg = TRUE
+    )
+  }
+
+  if (!file.exists(paste0(model_path, "/", cjson_filename))) {
+    stop(
+      "File \"",
+      cjson_filename,
+      "\" does not exist in model_path \"",
+      model_path,
+      "\"."
+    )
+  }
 
   if (is.null(sim_path)) sim_path <- model_path
 
@@ -335,10 +359,11 @@ write_config <- function(x,
 
     row_id <- NULL
[…]
