### 2023-12-05 7171303
add support for new cjson files with new arg to write_config - cjson_filename....
