### 2024-02-26 92301d6
Added convenience function get_cellindex
