+++ NAMESPACE
@@ -18,6 +18,7 @@ export(calc_cellarea)
 export(check_config)
 export(create_header)
 export(detect_io_type)
+export(get_cellindex)
 export(get_datatype)
 export(get_header_item)
 export(get_headersize)
