### 2024-02-26 0818318
Added tests and fixed some bugs
