### 2024-03-01 e3dc321
rework kit methods for LPJmLGridData, add subsetting for nearest neighbour for coordinates
