+++ R/LPJmLGridData.R
@@ -30,71 +30,6 @@ LPJmLGridData <- R6::R6Class( # nolint:object_name_linter
       stop("Not allowed for an object of class LPJmLGridData")
     },
 
-
-    #' @description
-    #' ! Not allowed to use dimension names of `LPJmLGridData$data`
-    #' array directly to subset each dimension to match the supplied vectors.
-    #'
-    #' @param ... See [`subset.LPJmLData()`]
-    subset = function(...) {
-
-      stop("Not allowed for an object of class LPJmLGridData")
-    },
-
-
-    #' @description
-    #' ! Not allowed to transform inner `LPJmLGridData$data` array
-    #' into another space or time format.
-    #'
-    #' @param ... See [`transform()`].
-    transform = function(...) {
-
-      stop("Not allowed for an object of class LPJmLGridData")
-    },
-
-    # export methods --------------------------------------------------------- #
-
-
-    #' @description
-    #' ! Not allowed to coerce (convert) an `LPJmLGridData` object into a
-    #' \link[raster]{raster} or \link[raster]{brick} object.
-    #'
-    #' @param ... See [`as_raster()`].
-    as_raster = function(...) {
-
-      stop("Not allowed for an object of class LPJmLGridData")
-    },
-
-
-    #' @description
-    #' ! Not allowed to coerce (convert) an `LPJmLGridData` object into a
-    #' \link[terra]{rast} object.
-    #'
-    #' @param ... See [`as_terra()`].
-    as_terra = function(...) {
-
-      stop("Not allowed for an object of class LPJmLGridData")
-    },
-
-
-    #' @description
-    #' ! Internal method only to be used for package development.
-    #'
-    #' @param ... See [`subset()`].
-    .__subset_space__ = function(...) {
-      private$.subset_space(...)
-    },
-
-
-    #' @description
-    #' ! Internal method only to be used for package development.
-    #'
-    #' @param ... See [`transform()`].
-    .__transform_space__ = function(...) {
-      private$.transform_space(...)
-    },
-
-
     # Create a new LPJmLGridData object only to be used internally or explicitly
     #' @description
     #' !Internal method only to be used for package development!
