+++ R/get_cellindex_2.R
@@ -1,0 +1,199 @@
+#' @title Get Cell Index
+#'
+#' @description This function returns the cell index from a grid file based on
+#' the provided extent or coordinates. If neither extent nor coordinates are
+#' provided, the full grid will be returned. If both extent and coordinates are
+#' provided, the function will stop and ask for only one of them. The extent
+#' should be a vector of length 4 in the form c(lonmin, lonmax, latmin, latmax).
+#' If the extent is not in the correct form, the function will swap the values
+#' to correct it.
+#'
+#' @param grid_filename A string representing the grid file name.
+#' @param extent A numeric vector of length 4 representing the extent
+#' (lonmin, lonmax, latmin, latmax).
+#' @param coordinates A list of two numeric vectors representing the coordinates.
+#'
+#'
+#' @return The cell index from the grid file based on the provided extent or
+#' coordinates.
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' get_cellindex(
+#'   grid_filename = "my_grid.bin.json",
+#'   extent = c(-123.25, -122.75, 49.25, 49.75)
+#' )
+#' get_cellindex(
+#'   grid_filename = "my_grid.bin.json",
+#'   coordinates = list(c(-123.25, -122.75), c(49.25, 49.75))
+#' )
+#' }
+#' @details
+#' The function reads a grid file specified by `grid_filename` and creates a
+#' data frame with columns for longitude, latitude, and cell number. The cell
+#' number is a sequence from 1 to the number of rows in the data frame.
+#'
+#' If an `extent` is provided, the function filters the cells to include only
+#' those within the specified longitude and latitude range. The `extent` should
+#' be a numeric vector of length 4 in the form c(lonmin, lonmax, latmin, latmax).
+#'
+#' If a list of `coordinates` is provided, the function filters the cells to
+#' include only those that match the specified coordinates. The `coordinates`
+#' should be a list of two numeric vectors representing the longitude and
+#' latitude values.
+#'
+#' If both `extent` and `coordinates` are provided, the function will stop and
+#' ask for only one of them. If neither `extent` nor `coordinates` are provided,
+#' the function will return the cell numbers for all cells in the grid.
+#'
+#' The function also includes checks for input types and values, and gives
+#' specific error messages for different error conditions. For example, it
+#' checks if the `grid_filename` exists, if the `extent` vector has the correct
+#' length, and if the `coordinates` list contains two vectors of equal length.
+get_cellindex <- function(grid_filename, extent = NULL, coordinates = NULL) {
+  # Check input types and values
+  check_filepath(grid_filename)
+  check_extent(extent)
+  # check_coordinates(coordinates)
+  check_extent_and_coordinates(extent, coordinates)
+  if (!is.null(extent)) {
+    extent <- correct_extent(extent)
+  }
+  check_coordinates_length(coordinates)
+
+  grid_lonlat <- read_io(filename = grid_filename) %>%
+    LPJmLGridData$new()
+
+  # Read the grid file and create a data frame
+  cells <- as.data.frame(grid_lonlat$data)
+  cells$cellindex <- as.numeric(row.names(cells)) + 1
+
+  # Get the range of longitude and latitude in the cells
+  lon_range <- range(cells$lon)
+  lat_range <- range(cells$lat)
+
+  # Check if extent values are within the longitude and latitude range in the cells
+  if (!is.null(extent)) {
[…]
