### 2024-03-01 97edbcb
list/tibble compatibility for subsetting via coordinates
