### 2024-03-14 dc29b6b
debug functionality and tests as well as  add tests
