### 2024-03-14 1dc1023
solve requests from stephen in review
