+++ R/LPJmLData_as.R
@@ -258,6 +258,7 @@ LPJmLData$set("private",
     # dimensions.
     if (data_subset$meta$._space_format_ == "lon_lat") {
 
+      # default handling of LPJmLData objects else inherited like LPJmLGridData
       if (class(data_subset)[1] == "LPJmLData") {
 
         data_subset$transform(to = "cell")
@@ -307,6 +308,7 @@ LPJmLData$set("private",
       } else {
         tmp_data <- data_subset$data
       }
+      # default handling of LPJmLData objects else inherited like LPJmLGridData
       if (class(data_subset)[1] == "LPJmLData") {
         tmp_raster[
           raster::cellFromXY(
@@ -408,6 +410,7 @@ LPJmLData$set("private",
     # dimensions.
     if (data_subset$meta$._space_format_ == "lon_lat") {
 
+      # default handling of LPJmLData objects else inherited like LPJmLGridData
       if (class(data_subset)[1] == "LPJmLData") {
 
         data_subset$transform(to = "cell")
@@ -475,6 +478,7 @@ LPJmLData$set("private",
       } else {
         tmp_data <- data_subset$data
       }
+      # default handling of LPJmLData objects else inherited like LPJmLGridData
       if (class(data_subset)[1] == "LPJmLData") {
         tmp_rast[
           terra::cellFromXY(
@@ -511,6 +515,7 @@ LPJmLData$set("private",
 
     # Support lazy loading of grid for meta files. This throws an error if no
     # suitable grid file is detected.
+    # default handling of LPJmLData objects else inherited like LPJmLGridData
     if (class(self)[1] == "LPJmLData") {
       self$add_grid()
     }
@@ -549,6 +554,7 @@ create_tmp_raster <- function(data_subset, is_terra = FALSE) {
 
   # Calculate grid extent from range to span raster
   if (data_subset$meta$._space_format_ == "cell") {
+    # default handling of LPJmLData objects else inherited like LPJmLGridData
     if (class(data_subset)[1] == "LPJmLData") {
       data_extent <- rbind(min = apply(data_subset$grid$data, "band", min),
                            max = apply(data_subset$grid$data, "band", max))
