### 2024-03-15 86c9eda
run buildLibrary and fix linter warnings
