+++ DESCRIPTION
@@ -54,4 +54,4 @@ Suggests:
     sf
 Config/testthat/edition: 3
 VignetteBuilder: knitr
-Date: 2024-02-27
+Date: 2024-03-15
