### 2024-03-15 e29978d
bugfix for subsetting via indices for time(year, month, day) dimensions
