+++ DESCRIPTION
@@ -18,7 +18,7 @@ Description: A collection of basic functions to facilitate the work with the
     configuration files.
 License: AGPL-3
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Roxygen: list(markdown = TRUE,
               r6 = TRUE)
 Encoding: UTF-8
@@ -54,4 +54,4 @@ Suggests:
     sf
 Config/testthat/edition: 3
 VignetteBuilder: knitr
-Date: 2023-12-05
+Date: 2024-03-15
