### 2024-03-15 557e1ea
bugfix for subsetting via indices for time(year, month, day) dimensions
