### 2024-03-25 589ad78
fix check for fmt in output files in LPJmL Runner
