### 2024-03-25 170dcfa
make compatible with new way of defining output formats via default_fmt, default_suffix
