### 2024-03-27 d5cbb9d
rework kit methods for LPJmLGridData, add subsetting for nearest neighbour for coordinates for better getcellindex matching
