+++ NAMESPACE
@@ -24,6 +24,7 @@ export(get_header_item)
 export(get_headersize)
 export(make_lpjml)
 export(read_config)
+export(read_grid)
 export(read_header)
 export(read_io)
 export(read_meta)

[…]
