+++ NAMESPACE
@@ -21,12 +21,14 @@ export(calc_cellarea)
 export(check_config)
 export(create_header)
 export(detect_io_type)
+export(get_cellindex)
 export(get_datatype)
 export(get_header_item)
 export(get_headersize)
 export(make_lpjml)
 export(plot.LPJmLData)
 export(read_config)
+export(read_grid)
 export(read_header)
 export(read_io)
 export(read_meta)

[…]
