+++ R/LPJmLData_as.R
@@ -318,7 +318,7 @@ LPJmLData$set(
         tmp_data <- data_subset$data
       }
       # default handling of LPJmLData objects else inherited like LPJmLGridData
-      if (class(data_subset)[1] == "LPJmLData") {
+      if (class(data_subset)[1] %in% c("LPJmLData", "LPJmLDataCalc")) {
         tmp_raster[
           raster::cellFromXY(
             tmp_raster,
@@ -566,7 +566,7 @@ create_tmp_raster <- function(data_subset, is_terra = FALSE) {
   # Calculate grid extent from range to span raster
   if (data_subset$meta$._space_format_ == "cell") {
     # default handling of LPJmLData objects else inherited like LPJmLGridData
-    if (class(data_subset)[1] == "LPJmLData") {
+    if (class(data_subset)[1] %in% c("LPJmLData", "LPJmLDataCalc")) {
       data_extent <- rbind(min = apply(data_subset$grid$data, "band", min),
                            max = apply(data_subset$grid$data, "band", max))
     } else {
