+++ R/LPJmLData_as.R
@@ -487,7 +487,7 @@ LPJmLData$set(
         tmp_data <- data_subset$data
       }
       # default handling of LPJmLData objects else inherited like LPJmLGridData
-      if (class(data_subset)[1] %in% c("LPJmLData", "LPJmLDataCalc")) {
+      if (class(data_subset)[1] == "LPJmLData") {
         tmp_rast[
           terra::cellFromXY(
             tmp_rast,
@@ -527,7 +527,7 @@ LPJmLData$set(
     # Support lazy loading of grid for meta files. This throws an error if no
     # suitable grid file is detected.
     # default handling of LPJmLData objects else inherited like LPJmLGridData
-    if (class(data_subset)[1] %in% c("LPJmLData", "LPJmLDataCalc")) {
+    if (class(self)[1] %in% c("LPJmLData", "LPJmLDataCalc")) {
       self$add_grid()
     }
 
