### 2024-04-08 c61ffd7
cast numeric to interger for subsetting
