### 2024-04-08 32ecb68
Added checks and changed tests
