### 2024-04-09 a1d63fe
Updated documentation of get_cellindex
