+++ man/get_cellindex.Rd
@@ -9,8 +9,8 @@ get_cellindex(grid_filename, extent = NULL, coordinates = NULL)
 \arguments{
 \item{grid_filename}{A string representing the grid file name.}
 
-\item{extent}{A numeric vector of length 4 representing the extent
-(lonmin, lonmax, latmin, latmax).}
+\item{extent}{A numeric vector (lonmin, lonmax, latmin, latmax) containing the
+longitude and latitude boundaries between which values included in the subset.}
 
 \item{coordinates}{A list of two named (lon, lat) numeric vectors representing the coordinates.}
 }
@@ -54,7 +54,7 @@ length, and if the \code{coordinates} list contains two vectors of equal length.
 \dontrun{
 get_cellindex(
   grid_filename = "my_grid.bin.json",
-  extent = c(-123.25, -122.75, 49.25, 49.75)
+  extent = c(-123.25, -122.75, 49.25, 49.75) # (lonmin, lonmax, latmin, latmax)
 )
 get_cellindex(
   grid_filename = "my_grid.bin.json",
