+++ DESCRIPTION
@@ -54,4 +54,4 @@ Suggests:
     sf
 Config/testthat/edition: 3
 VignetteBuilder: knitr
-Date: 2024-04-08
+Date: 2024-04-09
