### 2024-05-14 ca09f0f
try 2 search patterns in find_gridfile
