### 2024-05-14 593ed69
replace find_gridfile by more generic find_varfile
