+++ R/LPJmLData.R
@@ -39,7 +39,7 @@ LPJmLData <- R6::R6Class( # nolint:object_name_linter
         # If user has not supplied any parameters try to find a grid file in the
         # same directory as data. This throws an error if no suitable file is
         # found.
-        filename <- find_gridfile(private$.meta$._data_dir_)
+        filename <- find_varfile(private$.meta$._data_dir_, "grid")
 
         message(
           paste0(
@@ -446,88 +446,4 @@ aggregate_array <- function(x,
     }
   }
   data
-}
-
-
-#' Search for a grid file in a directory
-#'
-#' Function to search for a grid file in a specific directory.
-#'
-#' @param searchdir Directory where to look for a grid file.
-#' @param strict Boolean. If set to `TRUE`, grid file must be named "grid.*",
-#'   where "*" is one or two file extensions with 3 or 4 characters, e.g.
-#'   "grid.bin.json". If set to `FALSE`, the function will first try to match
-#'   the strict pattern. If unsuccessful, any filename that starts with "grid"
-#'   will be matched.
-#' @return Character string with the file name of a grid file upon success.
-#'   Function fails if no matching grid file can be detected.
-#'
-#' @details This function looks for file names in `searchdir` that match the
-#'   `pattern` parameter in its [`list.files()`] call. Files of type "meta" are
-#'   preferred. Files of type "clm" are also accepted. The function returns an
-#'   error if no suitable file or multiple files are found. Otherwise, the file
-#'   name of the grid file including the full path is returned.
-#' @noRd
-find_gridfile <- function(searchdir, strict = FALSE) {
-  # This will only match file names "grid.*", where * is one or two file
-  # extensions with 3 or 4 characters, e.g. "grid.bin" or "grid.bin.json".
-  grid_files <- list.files(
-    path = searchdir,
-    pattern = "^grid(\\.[[:alpha:]]{3,4})+$",
-    full.names = TRUE
-  )
-  if (length(grid_files) > 0) {
-    grid_types <- sapply(grid_files, detect_io_type) # nolint:undesirable_function_linter.
-    # Prefer "meta" file_type if present
-    if (length(which(grid_types == "meta")) == 1) {
-      filename <- grid_files[match("meta", grid_types)]
-    } else if (length(which(grid_types == "clm")) == 1) {
-      # Second priority "clm" file_type
-      filename <- grid_files[match("clm", grid_types)]
-    } else if (strict) {
-      # Stop if either multiple files per file type or not the right type have
-      # been detected
-      stop(
-        "Cannot detect grid file automatically.\n",
-        "$add_grid has to be called supplying parameters as for read_io."
-      )
-    }
-  } else if (strict) {
-    # Stop if no file name matching pattern detected
-    stop(
-      "Cannot detect grid file automatically.\n",
-      "$add_grid has to be called supplying parameters as for read_io."
-    )
-  } else {
-    # Less strict pattern matching any file name that starts with "grid*".
-    grid_files <- list.files(
-      path = searchdir,
-      pattern = "^grid",
-      full.names = TRUE
-    )
-    if (length(grid_files) > 0) {
-      grid_types <- sapply(grid_files, detect_io_type) # nolint:undesirable_function_linter.
-      # Prefer "meta" file_type if present
-      if (length(which(grid_types == "meta")) == 1) {
-        filename <- grid_files[match("meta", grid_types)]
-      } else if (length(which(grid_types == "clm")) == 1) {
-        # Second priority "clm" file_type
-        filename <- grid_files[match("clm", grid_types)]
-      } else {
-        # Stop if either multiple files per file type or not the right type have
-        # been detected
-        stop(
-          "Cannot detect grid file automatically.\n",
[…]
