### 2024-05-14 #5447 @marcosvital
Dear @mhesselbarth, would you be willing to review this submission for JOSS?

We carry out an open checklist-driven reviews here in GitHub issues, and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is an ongoing submission that already went through a first round of review, and we would like to have a second person to look into it. Let us know if you are available, ok?
