### 2024-05-14 4c1daf5
add missing variable to 2nd search pattern
