### 2024-05-16 c3eaf10
add tests for read_io for reservoirs
