### 2024-05-16 a29c8f0
update documentation and unit test for find_varfile
