### 2024-05-17 2cda40b
Replace find_gridfile() with more generic find_varfile() function
