+++ R/read_cdf.R
@@ -1,0 +1,250 @@
+# utility function to determine time resolution of a netcdf file
+get_timestep <- function(file_nc){
+  tunit <- file_nc$dim$time$units
+  tvals <- file_nc$dim$time$vals
+  if (grepl("years since", tunit, ignore.case = TRUE)){
+    tres <- "annual"
+    firstyr <- unlist(strsplit(unlist(strsplit(tunit, split = ' ', 
+                                               fixed = TRUE))[3], split = '-', fixed = TRUE))[1]
+  }else if (grepl("year", tunit, ignore.case = TRUE)){
+    tres <- "annual"
+    firstyr <- tvals[1]
+  }else if (grepl("days since", tunit, fixed = TRUE)){
+    ddiff <- tvals[2]-tvals[1]
+    firstyr <- unlist(strsplit(unlist(strsplit(tunit, split = ' ', 
+                                               fixed = TRUE))[3], split = '-', fixed = TRUE))[1]
+    if (ddiff > 27 && ddiff < 32){
+      tres <- "monthly"
+    }else if (ddiff > 364 && ddiff < 367){
+      tres <- "annual"
+    }else if (ddiff == 1){
+      tres <- "daily"
+    }else{
+      stop("Automatic detection of firstyear and time resolution failed.")
+    }
+  }else{
+    stop("Automatic detection of firstyear and time resolution failed.")
+  }
+  return(list(tres = tres, firstyr = as.numeric(firstyr)))
+}
+
+# utility function to guess the main variable of a netcdf file
+get_main_variable <- function(file_nc){
+  for (var in names(file_nc$var)){
+    ndims <- file_nc$var[[var]]$ndims
+    dimNames <- c()
+    for (d in 1:ndims){
+      dimNames <- append(dimNames,file_nc$var[[var]]$dim[[d]]$name) 
+    }
+    #print(paste0(var,paste(dimNames,collapse = ",")))
+    if (grepl("lon",paste(dimNames,collapse = " "),ignore.case = T) && grepl("lat",paste(dimNames,collapse = " "),ignore.case = T)){
+      return(var)
+    }
+  }
+  print(paste0("None of the variables could certainly be identified as main variable, guessing the last one: ",var))
+  return(var)
+}
+
+#' Reads netcdf and returns it header info
+#'
+#' Reads an arbitrary netcdf + returns the header values to read as lpjml object
+#'
+#' @param nc_in_file netcdf file name
+#' @param var optional variable to be read, in case automatic detection does 
+#'        not work as intended or several variables are stored within the file
+#' @param suppress_read_print whether to suppress the info "reading file XYZ" 
+#'        (default TRUE)
+#'
+#' @return header data
+#'
+#' @examples
+#' \dontrun{
+#' }
+#'
+#' @export
+read_cdf_header <- function( nc_in_file, 
+                          var = NULL,
+                          suppress_read_print = TRUE
+) {
+  file_type <- lpjmlkit::detect_io_type(filename = nc_in_file)
+  #read_cdf_meta <- lpjmlkit::read_meta(filename = fpc_file_meta)
+  empty_meta <- lpjmlkit::LPJmLMetaData
+  
+  varnames <- names(file_nc$var)
+  file_nc <- ncdf4::nc_open(filename = nc_in_file)
+  if (is.null(var)) var <- get_main_variable(file_nc = file_nc)
+  
+  # get the first of the variable names that has not been identified as the 
+  # main variable
+  bands_var_name <- varnames[-match(var,varnames)][1]
+  bands <- ncdf4::ncvar_get(nc = file_nc, varid = bands_var_name)
+                           
+  # get lon/lat information
[…]
