### 2024-06-03 5dc8fa7
continued read_cdf function
