+++ R/read_cdf.R
@@ -63,21 +63,26 @@ get_main_variable <- function(file_nc){
 #'
 #' @export
 read_cdf_header <- function( nc_in_file, 
-                          var = NULL,
-                          suppress_read_print = TRUE
+                             var = NULL,
+                             suppress_read_print = TRUE
 ) {
   file_type <- lpjmlkit::detect_io_type(filename = nc_in_file)
   #read_cdf_meta <- lpjmlkit::read_meta(filename = fpc_file_meta)
-  empty_meta <- lpjmlkit::LPJmLMetaData
-  
-  varnames <- names(file_nc$var)
+
   file_nc <- ncdf4::nc_open(filename = nc_in_file)
+  varnames <- names(file_nc$var)
   if (is.null(var)) var <- get_main_variable(file_nc = file_nc)
   
   # get the first of the variable names that has not been identified as the 
   # main variable
-  bands_var_name <- varnames[-match(var,varnames)][1]
-  bands <- ncdf4::ncvar_get(nc = file_nc, varid = bands_var_name)
+  if (length(varnames)>1){
+    bands_var_name <- varnames[-match(var,varnames)][1]
+    bands <- ncdf4::ncvar_get(nc = file_nc, varid = bands_var_name)
+  }else{
+    bands_var_name <- ""
+    bands <- 1
+  }
+
                            
   # get lon/lat information
   dimnames <- names(file_nc$dim)
@@ -87,6 +92,7 @@ read_cdf_header <- function( nc_in_file,
   lat <- file_nc$dim[[latdim]]$vals
   nlonin <- length(lon)
   nlatin <- length(lat)
+  
   # take the median difference between the cells as the resolution
   spatial_difference_lon <- lon[2:length(lon)] - lon[1:(length(lon)-1)]
   resolution_lon <- median(spatial_difference_lon)
@@ -99,21 +105,34 @@ read_cdf_header <- function( nc_in_file,
   timing <- get_timestep(file_nc = file_nc)
   tres <- timing$tres
   firstyr <- timing$firstyr
+  if (tres == "annual"){
+    nyears <- length(tvals)
+    nsteps <- 1
+  }else if (tres == "monthly"){
+    nyears <- length(tvals)/12
+    nsteps <- 12
+  }else {
+    stop("time resolution unknown. (e.g. daily not implemented yet)")
+  }
+  
+  lastyr <- firstyr + nyears - 1
   
   global_attributes <- ncdf4::ncatt_get(file_nc, 0)
   var_attributes <- ncdf4::ncatt_get(file_nc, var)
   
+  empty_meta <- list()
   empty_meta$sim_name <- global_attributes$title
-  empty_meta$source global_attributes$source
-  empty_meta$history global_attributes$history
-  empty_meta$variable <- tolower(var)
+  empty_meta$source <- global_attributes$source
+  empty_meta$history <- global_attributes$history
+  empty_meta$name <- tolower(var)
+  empty_meta$variable <- var
   empty_meta$descr <- var_attributes$longname
   empty_meta$unit <- var_attributes$units
   empty_meta$nbands <- length(bands)
   empty_meta$band_names <- bands
-  empty_meta$nyear <- length(tvals)
+  empty_meta$nyear <- nyears
   empty_meta$firstyear <- firstyr
-  empty_meta$lastyear <- tvals[length(tvals)]
+  empty_meta$lastyear <- lastyr
   empty_meta$ncell <- nlonin*nlatin
   empty_meta$firstcell <- 0
   empty_meta$cellsize_lon <- resolution_lon
@@ -121,36 +140,38 @@ read_cdf_header <- function( nc_in_file,
   empty_meta$scalar <- 1
[…]
