### 2024-06-04 8dd4172
continued 2 read_cdf function
