### 2024-06-04 f8dee1c
continued 3 work on read_cdf function
