+++ R/read_cdf.R
@@ -66,7 +66,7 @@ get_main_variable <- function(file_nc){
 #' @export
 read_cdf_header <- function( nc_in_file, 
                              var = NULL,
-                             suppress_read_print = TRUE
+                             silent = TRUE
 ) {
   file_type <- lpjmlkit::detect_io_type(filename = nc_in_file)
   #read_cdf_meta <- lpjmlkit::read_meta(filename = fpc_file_meta)
@@ -146,37 +146,35 @@ read_cdf_header <- function( nc_in_file,
   empty_meta$format <- file_type
   empty_meta$filename <- basename(nc_in_file)
   empty_meta$nstep <- nsteps
-  
-  empty_meta$timestep <- 1
   empty_meta$datatype <- "float"
-  empty_meta$order <- "cellseq"
-  empty_meta$bigendian <- FALSE
+  empty_meta$order <- ""
   empty_meta$subset <- FALSE
+  
+  empty_meta$timestep <- 1 # todo: deal with this type
+  empty_meta$bigendian <- FALSE # ?
   ncdf4::nc_close(file_nc)
   
   header_data <- lpjmlkit::LPJmLMetaData$new(x = empty_meta)
-  return(header_data)
+  header_data
 }
 
 #' Reads netcdf and returns it as array
 #'
 #' Reads an arbitrary netcdf and returns the values at the lon/lat locations..
 #' Todos:
-#' - annual data which is written out only every X years (what about monthly?):
+#' - annual data which is written out only every X years:
 #'   (/p/tmp/sibylls/Methane/output_holocene/soilc.nc)
 #' - daily timesteps
 #' - check calendar
 #' - check lpj code for nc output creation
 #' - how to distribute nlat and nlon from nc header to read function? 
 #'   or read in again?
+#' - subsetting cells does not work yet
 #'
 #' @param nc_in_file netcdf file name
 #' @param nc_header header data, read in from either meta file or data in 
 #'        netcdf file
-#' @param get_year_start first year to be read 
-#'        (if not specified will default to first record year)
-#' @param get_year_stop final year to be read 
-#'        (if not specified will default to last record year)
+#' @param subset list object defining which subset of the data to be read
 #' @param suppress_read_print whether to suppress the info "reading file XYZ" 
 #'        (default TRUE)
 #'
@@ -189,9 +187,8 @@ read_cdf_header <- function( nc_in_file,
 #' @export
 read_cdf <- function( nc_in_file, 
                       nc_header,
-                      get_year_start = NULL,
-                      get_year_stop = NULL,
-                      suppress_read_print = TRUE
+                      subset = list(),
+                      silent = TRUE
 ) {
   var <- nc_header$variable
   file_nc <- ncdf4::nc_open(filename = nc_in_file)
@@ -202,9 +199,33 @@ read_cdf <- function( nc_in_file,
                  " please specify manually via argument var."))
   }
 
-  if (is.null(get_year_start)) get_year_start <- nc_header$firstyear
-  if (is.null(get_year_stop)) get_year_stop <- nc_header$lastyear
-  ngetyears <- get_year_stop - get_year_start + 1
+  # Determine all years in the file
+  years <- seq(
+    from       = default(nc_header$firstyear, 1901),
+    by         = default(nc_header$timestep, 1),
+    length.out = default(nc_header$nyear, 1)
+  )
+  # Years to read
+  if ("year" %in% names(subset)) {
+    if (is.numeric(subset[["year"]])) {
+      years <- years[subset[["year"]]]
+    } else {
+      years <- as.integer(subset[["year"]])
[…]
