+++ R/read_io.R
@@ -320,82 +320,107 @@ read_io <- function( # nolint:cyclocomp_linter.
   # warnings should have been triggered in read_io_metadata already.
   file_header <- meta_data$as_header(silent = TRUE)
 
-  # Check file size
-  # Check if file is an LPJDAMS input file, which has a different format.
-  if (get_header_item(file_header, "name") == "LPJDAMS") {
-    # Hardcoded size of 4 and number of bands
-    expected_filesize <- unname(
-      get_header_item(file_header, "ncell") *
-        length(band_names_reservoir) *
-        get_header_item(file_header, "nstep") *
-        get_header_item(file_header, "nyear") *
-        4 + start_offset
-    )
-  } else {
-    expected_filesize <- unname(
-      get_header_item(file_header, "ncell") *
-        get_header_item(file_header, "nbands") *
-        get_header_item(file_header, "nstep") *
-        get_header_item(file_header, "nyear") *
-        get_datatype(file_header)$size + start_offset
-    )
-  }
-  if (file.size(filename) != expected_filesize) {
-    stop(
-      "Unexpected file size (", file.size(filename), " bytes) of ", filename,
-      "\nExpected size: ", expected_filesize, " bytes",
-      "\nPlease check ",
-      switch(file_type, meta = "meta file", clm = "header", "supplied"),
-      " attributes."
-    )
-  }
+  if (!meta_data$format == "cdf"){
+    # Check file size
+    # Check if file is an LPJDAMS input file, which has a different format.
+    if (get_header_item(file_header, "name") == "LPJDAMS") {
+      # Hardcoded size of 4 and number of bands
+      expected_filesize <- unname(
+        get_header_item(file_header, "ncell") *
+          length(band_names_reservoir) *
+          get_header_item(file_header, "nstep") *
+          get_header_item(file_header, "nyear") *
+          4 + start_offset
+      )
+    } else {
+      expected_filesize <- unname(
+        get_header_item(file_header, "ncell") *
+          get_header_item(file_header, "nbands") *
+          get_header_item(file_header, "nstep") *
+          get_header_item(file_header, "nyear") *
+          get_datatype(file_header)$size + start_offset
+      )
+    }
+    if (file.size(filename) != expected_filesize) {
+      stop(
+        "Unexpected file size (", file.size(filename), " bytes) of ", filename,
+        "\nExpected size: ", expected_filesize, " bytes",
+        "\nPlease check ",
+        switch(file_type, meta = "meta file", clm = "header", "supplied"),
+        " attributes."
+      )
+    }
+  
+    # Check whether nbands may actually be nstep
+    if (!silent && get_header_item(file_header, "version") < 4 &&
+          get_header_item(file_header, "nstep") == 1 &&
+          get_header_item(file_header, "nbands") %in% c(12, 365)) {
+      message(
+        "read_io: Detected \"nbands = ", get_header_item(file_header, "nbands"),
+        "\" and \"nstep = 1\". If this is a ",
+        ifelse(get_header_item(file_header, "nbands") == 12, "monthly", "daily"),
+        " file consider setting \"nbands = 1\" and \"nstep = ",
+        get_header_item(file_header, "nbands"),
+        "\" to allow correct setting of the time axis."
+      )
+    }
+    # Read data from binary file
+    if (get_header_item(file_header, "name") == "LPJDAMS") {
+      file_data <- read_io_reservoir(filename, meta_data, subset, silent)
+    } else {
+      file_data <- read_io_data(filename, meta_data, subset, silent)
+    }
+    # Update meta_data based on subset
+    if (!is.null(subset$year) && is.numeric(subset$year)) {
[…]
