### 2024-06-04 f3bf66d
updated test_cdr.R
