### 2024-06-05 6d5a7d8
fixed timestepping and some more todo's
