+++ R/read_cdf.R
@@ -2,13 +2,17 @@
 get_timestep <- function(file_nc){
   tunit <- file_nc$dim$time$units
   tvals <- file_nc$dim$time$vals
+  timestep <- 1 # default
   if (grepl("years since", tunit, ignore.case = TRUE)){
     tres <- "annual"
-    firstyr <- unlist(strsplit(unlist(strsplit(tunit, split = ' ', 
+    baseyr <- unlist(strsplit(unlist(strsplit(tunit, split = ' ', 
                               fixed = TRUE))[3], split = '-', fixed = TRUE))[1]
+    offset_yr <- tvals[1]
+    firstyr <- baseyr + offset_yr
   }else if (grepl("year", tunit, ignore.case = TRUE)){
     tres <- "annual"
     firstyr <- tvals[1]
+    timestep <- tvals[2]-tvals[1]
   }else if (grepl("days since", tunit, fixed = TRUE)){
     ddiff <- tvals[2]-tvals[1]
     baseyr <- as.integer(unlist(strsplit(unlist(strsplit(tunit, split = ' ', 
@@ -27,7 +31,7 @@ get_timestep <- function(file_nc){
   }else{
     stop("Automatic detection of firstyear and time resolution failed.")
   }
-  return(list(tres = tres, firstyr = as.numeric(firstyr)))
+  return(list(tres = tres, firstyr = as.numeric(firstyr), timestep = timestep))
 }
 
 # utility function to guess the main variable of a netcdf file
@@ -50,11 +54,13 @@ get_main_variable <- function(file_nc){
 #' Reads netcdf and returns it header info
 #'
 #' Reads an arbitrary netcdf + returns the header values to read as lpjml object
+#' Todos:
+#' - check calendar for irregularities (leapdays)?
 #'
 #' @param nc_in_file netcdf file name
 #' @param var optional variable to be read, in case automatic detection does 
 #'        not work as intended or several variables are stored within the file
-#' @param suppress_read_print whether to suppress the info "reading file XYZ" 
+#' @param silent whether to suppress the info "reading file XYZ" 
 #'        (default TRUE)
 #'
 #' @return header data
@@ -106,7 +112,8 @@ read_cdf_header <- function( nc_in_file,
   timing <- get_timestep(file_nc = file_nc)
   tres <- timing$tres
   firstyr <- timing$firstyr
-  #browser()
+  timestep <- timing$timestep
+
   if (tres == "annual"){
     nyears <- length(tvals)
     nsteps <- 1
@@ -119,39 +126,40 @@ read_cdf_header <- function( nc_in_file,
     if (!all.equal(round(nyears,0), nyears)) 
        stop("Number of time records not a multiple of 365 - pls check calendar")
   }
-  
-  lastyr <- firstyr + nyears - 1
-  
+  if (timestep == 1) lastyr <- firstyr + nyears - 1
+  else lastyr <- tvals[length(tvals)]
+    
   global_attributes <- ncdf4::ncatt_get(file_nc, 0)
   var_attributes <- ncdf4::ncatt_get(file_nc, var)
-  
+
   empty_meta <- list()
   empty_meta$sim_name <- global_attributes$title
   empty_meta$source <- global_attributes$source
   empty_meta$history <- global_attributes$history
   empty_meta$name <- tolower(var)
   empty_meta$variable <- var
-  empty_meta$descr <- var_attributes$longname
+  empty_meta$descr <- var_attributes$long_name
   empty_meta$unit <- var_attributes$units
   empty_meta$nbands <- length(bands)
   empty_meta$band_names <- bands
   empty_meta$nyear <- nyears
   empty_meta$firstyear <- firstyr
   empty_meta$lastyear <- lastyr
[…]
