### 2024-06-17 1be812a
added lpj_cmd argument and fixed linter warnings
