### 2024-06-17 03ea940
refactor lpj_cmd and update documentation
