+++ R/run_lpjml.R
@@ -19,9 +19,9 @@
 #'   written, including output, restart and configuration files. If `NULL`,
 #'   `model_path` is used. See also [write_config]
 #'
-#' @param lpj_cmd Character string defining the command to run LPJmL.
-#'   Defaults to "srun --propagate" (old cluster at PIK).
-#'   Change to "mpirun" for HPC2024 at PIK.
+#' @param run_cmd Character string defining the command used to execute lpjml
+#'   (see details). Defaults to "srun --propagate" (compute ondes of old cluster
+#'   at PIK). Change to "mpirun" for HPC2024 at PIK.
 #'
 #' @param parallel_cores Integer defining the number of available CPU
 #'   cores/nodes for parallelization. Defaults to `1` (no parallelization).
@@ -41,6 +41,8 @@
 #' @return See `x`, extended by columns `"type"`, `"job_id"` and `"status"`.
 #'
 #' @details
+#'
+#' \bold{x}:
 #' A \link[tibble]{tibble} for `x` that has been generated by
 #' [`write_config()`] and can look like the following examples can
 #' supplied:
@@ -65,6 +67,18 @@
 #' character string vector directly as the `x` argument to `run_lpjml`. \cr
 #' Also be aware that the order of the supplied config files is important
 #' (e.g. make sure the spin-up run is run before the transient one).
+#'
+#' \bold{run_cmd}:
+#' The `run_cmd` argument is used to define the command to execute LPJmL. This
+#' is needed because the LPJmL executable can not directly be used on all
+#' machines. Which command has to be used depends on the software installed.
+#' Further information on this can be found in the INSTALL file of LPJmL.
+#' To determine the correct command, check the lpj_submit.sh file in the bin
+#' directory of LPJmL. Using PIK infrastrucure the command is \code{srun} for
+#' the hpc2015 and \code{mpirun} for the hpc2024. To facilitate usage on the
+#' interactive (login) nodes, no command is needed for hpc2015. For the hpc2024
+#' the command remains \code{mpirun} (in these cases \code{run_lpjml} adjusts
+#' \code{run_cmd} accordingly).
 #'
 #' @examples
 #'
@@ -153,7 +167,7 @@
 run_lpjml <- function(x,
                       model_path = ".",
                       sim_path = NULL,
-                      lpj_cmd = "srun --propagate",
+                      run_cmd = "srun --propagate",
                       parallel_cores = 1,
                       write_stdout = FALSE,
                       raise_error = TRUE,
@@ -196,11 +210,11 @@ run_lpjml <- function(x,
 
       if (parallel_cores == 1) {
         do_sequential(
-          sim_names, model_path, sim_path, lpj_cmd, write_stdout, raise_error
+          sim_names, model_path, sim_path, run_cmd, write_stdout, raise_error
         )
       } else if (parallel_cores > 1 && Sys.getenv("SLURM_JOB_ID") != "") {
         do_parallel(
-          sim_names, model_path, sim_path, lpj_cmd, parallel_cores, raise_error
+          sim_names, model_path, sim_path, run_cmd, parallel_cores, raise_error
         )
       } else {
         stop(
@@ -214,11 +228,11 @@ run_lpjml <- function(x,
   } else {
     if (parallel_cores == 1) {
       do_sequential(
-        x$sim_name, model_path, sim_path, lpj_cmd, write_stdout, raise_error
+        x$sim_name, model_path, sim_path, run_cmd, write_stdout, raise_error
       )
     } else if (parallel_cores > 1 && Sys.getenv("SLURM_JOB_ID") != "") {
       do_parallel(
-        x$sim_name, model_path, sim_path, lpj_cmd, parallel_cores, raise_error
[…]
