+++ man/run_lpjml.Rd
@@ -8,6 +8,7 @@ run_lpjml(
   x,
   model_path = ".",
   sim_path = NULL,
+  run_cmd = "srun --propagate",
   parallel_cores = 1,
   write_stdout = FALSE,
   raise_error = TRUE,
@@ -28,6 +29,10 @@ multiple generated configuration file(s).}
 \item{sim_path}{Character string defining path where all simulation data are
 written, including output, restart and configuration files. If \code{NULL},
 \code{model_path} is used. See also \link{write_config}}
+
+\item{run_cmd}{Character string defining the command used to execute lpjml
+(see details). Defaults to "srun --propagate" (compute ondes of old cluster
+at PIK). Change to "mpirun" for HPC2024 at PIK.}
 
 \item{parallel_cores}{Integer defining the number of available CPU
 cores/nodes for parallelization. Defaults to \code{1} (no parallelization).
@@ -54,6 +59,7 @@ that can be used as an input (see \code{x}). It contains the details to run sing
 or multiple (dependent/subsequent) model runs.
 }
 \details{
+\bold{x}:
 A \link[tibble]{tibble} for \code{x} that has been generated by
 \code{\link[=write_config]{write_config()}} and can look like the following examples can
 supplied:\tabular{l}{
@@ -77,6 +83,18 @@ As a shortcut it is also possible to provide the config file
 character string vector directly as the \code{x} argument to \code{run_lpjml}. \cr
 Also be aware that the order of the supplied config files is important
 (e.g. make sure the spin-up run is run before the transient one).
+
+\bold{run_cmd}:
+The \code{run_cmd} argument is used to define the command to execute LPJmL. This
+is needed because the LPJmL executable can not directly be used on all
+machines. Which command has to be used depends on the software installed.
+Further information on this can be found in the INSTALL file of LPJmL.
+To determine the correct command, check the lpj_submit.sh file in the bin
+directory of LPJmL. Using PIK infrastrucure the command is \code{srun} for
+the hpc2015 and \code{mpirun} for the hpc2024. To facilitate usage on the
+interactive (login) nodes, no command is needed for hpc2015. For the hpc2024
+the command remains \code{mpirun} (in these cases \code{run_lpjml} adjusts
+\code{run_cmd} accordingly).
 }
 \examples{
 
