### 2024-06-23 #5447 @arfon
@marcosvital @kthyng – incase it's useful, I've generated the most similar paper comparison here: https://github.com/openjournals/joss-reviews/issues/5314#issuecomment-2184969892
