### 2024-07-16 PIK-LPJmL/lpjmlkit#1 @kanishkan91
JOSS review: Add GitHub installation instructions
Authors have added detailed installation instructions for the CRAN version of the package. However, Since this package is available on GitHub, I would recommend the authors also add instructions on how users can directly install the package from GitHub.

I was able to install the same using- `devtools::install_github("PIK-LPJmL/lpjmlkit",build_vignettes = TRUE)`

I would recommend authors add instructions for the same in the README.

https://github.com/openjournals/joss-reviews/issues/5447
