### 2024-07-16 PIK-LPJmL/lpjmlkit#2 @kanishkan91
JOSS review: Generate docs using pkgdown
Since the repository is already set up very nicely, I recommend that the authors generate the documentation for this software package as html docs using `pkgdown`. Essentially all that would be required would be to install `pkgdown` and run `pkgdown::build_site()`. This would convert all documentation (e.g. functions, vignettes) to html docs. The same can be hosted on GitHub docs and authors can even add a link to the same in the "About" section

https://github.com/openjournals/joss-reviews/issues/5447
