### 2024-07-16 PIK-LPJmL/lpjmlkit#3 @kanishkan91
JOSS review: Comments on documentation, README
- [ ] In the contribution section, please indicate that users are encouraged to report issues on GitHub
- [ ] I would recommend removing the "citation" since once the JOSS paper is published, you will just add a button on the README

https://github.com/openjournals/joss-reviews/issues/5447
