### 2024-07-16 PIK-LPJmL/lpjmlkit#4 @kanishkan91
JOSS review: Unable to test functions in vignettes since "outputs" folder is not available 
When I try to run `runoff <- read_io("./output/mrunoff.bin.json")` it does not work since the output folder does not exist. It does exist within the tests folder. This may confuse new users. I would recommend adding a sample_data folder or the outputs folder itself to the root directory to resolve this. Or maybe mention that the sample data needs to be downloaded or something?

https://github.com/openjournals/joss-reviews/issues/5447
