### 2024-07-20 PIK-LPJmL/lpjmlkit#5 @kanishkan91
JOSS review: Can authors add more explnations regarding the the LpJML installation required to run this software package?
Perhaps I am missing something fundamental here, but since this package requires a pre-requisite installation of LpJML, can the authors describe how the same can be installed? I agree that the model itself is probably well documented already, however can the authors link to that documenation? Particularly this may be necessary to mention in the vignettes.

Also, how are the tests in the current package solving without a pre-installed LpJML?

https://github.com/openjournals/joss-reviews/issues/5447
