### 2024-07-20 #5447 @kanishkan91
@marcosvital , @jnnsbrr - I just completed my review of this software package. I was delayed on account of deadlines and paternity leave, so my apologies in advance. I have added my questions as issues which the authors can close with relevant pull requests. Most of the comments should be minor and easy to respond to. The two that would require a bit more clarification are the issues with the vignettes and the explanations regarding pre-installing lpjml. Once the authors respond, I am happy to recommend publication.
