### 2024-07-26 e6e3849
docs compatible with buildLibrary
