### 2024-07-26 1a5e498
add lpjml installation sentence in vignette
