+++ vignettes/lpjml-data.Rmd
@@ -28,7 +28,8 @@ module that groups around the data class `LPJmLData` and aims to facilitate
 reading and processing of LPJmL inputs and outputs by combining the raw data
 with available meta data (meta files, header files or manually) to avoid a large
 overhead. It further enables easy subsetting, transformations and basic
-statistics of the data and allows export to common data formats.
+statistics of the data and allows export to common data formats.  
+Example data files can be downloaded from https://doi.org/10.5281/zenodo.12915168
 
 &nbsp;
 
@@ -50,7 +51,7 @@ or exported (5) to common data formats.
   in "meta" format.
   ```{r, echo = TRUE, eval = FALSE, highlight = TRUE}
   # Read monthly runoff with meta file.
-  runoff <- read_io("./output/mrunoff.bin.json")
+  runoff <- read_io("./output/runoff.bin.json")
   ```
 
 * **"clm"** - *Use if "meta" is not available or in combination*.  
@@ -60,16 +61,16 @@ or exported (5) to common data formats.
   information about file structure is derived automatically from the file header.
   ```{r, echo = TRUE, eval = FALSE, highlight = TRUE}
   # Read monthly runoff data with header.
-  runoff <- read_io("./output/mrunoff.clm",
-                   # If the clm version is lower than 4 set nbands and nstep
-                   # manually so that month dimension is recognized correctly.
-                   nbands = 1,
-                   nstep = 12,
-                   # Useful additional information that is not needed to read the
-                   # Data.
-                   variable = "runoff",
-                   descr = "monthly runoff",
-                   unit = "mm/month")
+  runoff <- read_io("./output/runoff.clm",
+                    # If the clm version is lower than 4 set nbands and nstep
+                    # manually so that month dimension is recognized correctly.
+                    nbands = 1,
+                    nstep = 12,
+                    # Useful additional information that is not needed to read the
+                    # Data.
+                    variable = "runoff",
+                    descr = "monthly runoff",
+                    unit = "mm/month")
   ```
 
 * **"raw"** - *Not recommended for use (with lpjmlkit)*.  
@@ -81,10 +82,10 @@ or exported (5) to common data formats.
   LPJmL input files use "raw" format.
   ```{r, echo = TRUE, eval = FALSE, highlight = TRUE}
   # Read monthly runoff raw data.
-  runoff <- read_io("./output/mrunoff.bin",
-                   # Specify all meta data if they differ from the function
-                   # default values.
-                   ...)
+  runoff <- read_io("./output/runoff.bin",
+                    # Specify all meta data if they differ from the function
+                    # default values.
+                    ...)
   ```
 
 &nbsp;
@@ -195,14 +196,14 @@ plot(runoff)
   **$grid** attribute (as an LPJmLData object) to the object, providing the spatial
   reference (longitude and latitude) for each cell.
   ```{r, echo = TRUE, eval = FALSE, highlight = TRUE}
[…]
