### 2024-07-26 efe3fc6
merge master (including latest issue fixes) and thank david hoetten in acknowledgements
