### 2024-07-26 PIK-LPJmL/lpjmlkit#5 @jnnsbrr
Dear @kanishkan91 LPJmL is only required for the LPJmLRunner part, I added a line including a link to the model INSTALL file in the start of the [LPJmLRunner Vignette](https://github.com/PIK-LPJmL/lpjmlkit/blob/051c06930a658347c72765cc03502cde17ec0bc6/vignettes/lpjml-runner.Rmd#L31)

> To install LPJmL, read the [LPJmL installation instructions](https://github.com/PIK-LPJmL/LPJmL/blob/master/INSTALL).
