### 2024-07-26 PIK-LPJmL/lpjmlkit#4 @jnnsbrr
Thanks @kanishkan91. Unfortunately LPJmL ouput data is quite large in size. Therefore I created a zenodo publication with some example data that can be used - which is linked to at the [Start of the LPJmLData vignette](https://github.com/PIK-LPJmL/lpjmlkit/blob/051c06930a658347c72765cc03502cde17ec0bc6/vignettes/lpjml-data.Rmd#L32) where the data is useful.
> Example data files can be downloaded from https://doi.org/10.5281/zenodo.12915168
