### 2024-07-26 PIK-LPJmL/lpjmlkit#3 @jnnsbrr
Dear @kanishkan91, unfortunately parts of the README are automatically build and standardized (including citation and contributing) using an R library at our Institute (https://github.com/pik-piam/lucode2) which follows a common structure (https://github.com/orgs/pik-piam/repositories) I cannot directly change.
But I saw that the CONTRIBUTING.md file was missing, so its now included and provides detailed instructions.
