### 2024-07-26 PIK-LPJmL/lpjmlkit#2 @jnnsbrr
[Done.](https://github.com/PIK-LPJmL/lpjmlkit/tree/master/docs)
