### 2024-07-26 PIK-LPJmL/lpjmlkit#1 @jnnsbrr
As described in issue #3 this is also part of an automatically build README which is why I cannot change it in the installation part :-/
