### 2024-07-26 #5447 @jnnsbrr
Dear @kanishkan91 thank you very much for review!
I tried to solve the [issues you created](https://github.com/PIK-LPJmL/lpjmlkit/issues) as requested - still parts of our README.md is automatically build [by an R library that standardizes R package development at PIK](https://github.com/pik-piam/lucode2).
I hope that my changes still solve them :-)
