### 2024-07-26 PIK-LPJmL/lpjmlkit#5 @kanishkan91
Thanks. Though, regarding the question above, how are the tests passing on GitHub actions if an installation of LpJML is required? This won't hold the review back. I was just curious.
