### 2024-07-26 #5447 @kanishkan91
@jnnsbrr Thanks for your responses. I think this should be good to go. Though I had two minor questions. I just want this to be documented in the review thread.

1. If an LpJML installation is a pre-requisite for this package, how are the tests solving on GitHub actions? Is a pre-installed LpJML added somewhere? Or is LpJML being simultaneously installed by Github actions Because the tests are covering 90% of the code base, I find this a bit intriguing.
2. Regarding the inability to edit the README, to be honest that is a bit unusual. I still approve publication, but can you comment on why you cant edit the README. Usually, there are pre-compiled templates that teams work with but at the end of the day, the README being a fundamental document should be able to be edited? I would also recommend you keep the above "issues" on the README open on your repo, just in case someone else had the same questions.

@marcosvital - I finished up the checklists.
