### 2024-07-26 PIK-LPJmL/lpjmlkit#5 @jnnsbrr
Ah sorry, forgot to answer on that.
No this part of the function is skipped since that would require a test instance to run LPJmL which requires a more complex setup. And since the package does not only features the LPJmLRunner but functions to work with the data on very different machines we cannot provide such a setup. But we aimed to cover as much as possible of these functions with tests :-), see for example the test coverage of [`run_lpjml()`](https://app.codecov.io/gh/PIK-LPJmL/lpjmlkit/blob/master/R%2Frun_lpjml.R) or [`submit_lpjml()`](https://app.codecov.io/gh/PIK-LPJmL/lpjmlkit/blob/master/R%2Fsubmit_lpjml.R).
