### 2024-07-26 #5447 @jnnsbrr
Dear @kanishkan91 thanks again!
On 1. I answered already to some extent in [the issue](https://github.com/PIK-LPJmL/lpjmlkit/issues/5). There is currently no way on running an LPJmL instance within GitHub actions. There might a way for that in the future but currently this is not being worked on. We can still cover as much as possible of the lpjmlkit functionality with tests to make sure there are no bugs/failures on our side of the pipeline.

On 2: Yes you are right, its a bit unusual. Still the RSE Team at PIK provides us with a lot of services, distribution on internal servers, Runiverse and more (https://pik-piam.r-universe.dev/). The pik-piam team has introduced this workflow for a lot of R packages and its quite an effort for us to branch off here. But I will try to open an issue on that soon to make it more flexible. This is the [corresponding function](https://pik-piam.r-universe.dev/lucode2/doc/manual.html#buildLibrary)
