### 2024-09-23 fed341d
prevent invalid LPJmLGridData object if variable is missing in metadata
