### 2024-09-24 db9b115
Added shape file support and changed return value of get_cellindex
