### 2024-10-03 c447dbd
adjust draft_pdf.yml
