### 2024-10-03 #5447 @jnnsbrr
Dear @kthyng,

thanks a lot. We went trough the list and released the latest version (v1.7.3) that is also now featured in the paper. Here is the Zenodo link: https://doi.org/10.5281/zenodo.13884611.
Author information do match with an additional contributor that is now part as author.
