### 2024-10-22 4c6f7a9
Update paper.md
Proposed edits
