+++ paper.md
@@ -65,7 +65,7 @@ developed for handling the open source dynamic global vegetation model (DGVM)
 It contains two main modules.
 One, *LPJmL Runner*, provides the functionality to create multiple model
 configurations and start the corresponding simulations either on a personal
-computer or on an HPC (High Perfomance Computing) cluster with SLURM
+computer or on a high perfomance computing a (HPC) cluster with SLURM
 (Simple Linux Utility for Resource Management) support, in both cases requiring
 a working LPJmL installation.
 The other, *LPJmL Data*, offers a generic function that supports reading both
@@ -90,21 +90,21 @@ reusable, according to the FAIR principles for research software
 [@barker_introducing_2022].
 Numerical models based on process-based approaches and implemented in low-level
 programming languages are often characterised by a long history and long
-development cycles, that affect operability.
+development cycles that affect operability.
 These models therefore often lack simple and reliable interfaces that are also
 FAIR [@wilson_best_2014; @barker_introducing_2022].
-LPJmL is a well-established dynamic global vegetation model (DGVM), widely used
+LPJmL is a well-established dynamic global vegetation model (DGVM) that is widely used
 in the scientific community.
 A DGVM simulates potential vegetation shifts and associated biogeochemical and
-hydrological cycles in response to changes in climate. [@cramer_global_2001]
+hydrological cycles in response to changes in climate [@cramer_global_2001].
 DGVMs are typically used to simulate the effects of future climate change on
 natural vegetation and its carbon and water cycles.
 In this context LPJmL has been used for more than a decade and was employed by
 researchers to conduct numerous studies in various research areas related to the
 terrestrial biosphere.
 To this end, the original DGVM LPJ [@Sitch2003] was extended by adding an
-improved representation of the hydrological cycle [@gerten_terrestrial_2004],
-by implementing managed land components, forming "LPJmL"
+improved representation of the hydrological cycle [@gerten_terrestrial_2004]. 
+This was accomplished by implementing managed land components, forming "LPJmL"
 (LPJ with *m*anaged *L*and)) [@bondeau_modelling_2007; @rolinski_modeling_2018;
 @lutz2019simulating; @schaphoff_lpjml4_2018], and by including the nitrogen 
 cycle [@von_bloh_implementing_2018].
@@ -169,7 +169,7 @@ e.g. multi-scenario or uncertainty/sensitivity analyses.
 The creation of these tables is based on the concept of `tidy data` with columns as
 parameters/settings (variables) and rows as simulations (observations)
 [@wickham_tidy_2014].
-In order to facitlitate the execution of these simulations, the function `write_config()`
+In order to facilitate the execution of these simulations, the function `write_config()`
 handles the formatting of the corresponding `tidy data` tables into LPJmL compatible JSON
 configuration files.
 
@@ -192,7 +192,7 @@ Using `read_io()`, LPJmL raw output data can be read into a standardised data fo
 `LPJmLData`, which makes use of the corresponding metadata
 to correctly represent the data arrays (dimensions, units, etc.).
 
-This data format is designed to faciliitate the functionalities commonly used by
+This data format is designed to facilitate the functionalities commonly used by
 LPJmL users for data analysis.
 In addition to the descriptive statistics that an LPJmLData object displays by default,
 there is a `plot()` method to easily display the the data for visual inspection.
@@ -220,7 +220,7 @@ aggregation over years and months allows the following presentation.
 ![Exemplary LPJmLData plot of variable runoff generated by functions read_io, transform, subset and plot](inst/img/example_plot.png)
 
 
-## Miscellaneos
+## Miscellaneous
 
 The *lpjmlkit* package also includes additional functions to support different
 applications that involve handling LPJmL or related data.
