### 2024-10-22 #5447 @kthyng
proposed paper edits: https://github.com/PIK-LPJmL/lpjmlkit/pull/6
