### 2024-11-15 5baae9d
enable negative years in split_time_names
