### 2024-11-15 e9d56d5
fix linter warnings, build library
