### 2024-11-18 #5447 @jnnsbrr
Dear @kthyng, thanks a lot!
So if I see it correctly the missing DOIs are the only problem that remains?
For the r6 package it seems that they do not provide a DOI, [this is how they recommend to cite them](https://r6.r-lib.org/authors.html#citation:~:text=%40Manual%7B%2C%0A%20%20title%20%3D%20%7BR6%3A%20Encapsulated%20Classes%20with%20Reference%20Semantics%7D%2C%0A%20%20author%20%3D%20%7BWinston%20Chang%7D%2C%0A%20%20year%20%3D%20%7B2022%7D%2C%0A%20%20note%20%3D%20%7Bhttps%3A//r6.r%2Dlib.org%2C%20https%3A//github.com/r%2Dlib/R6/%7D%2C%0A%7D). Would that be ok?
