### 2024-11-23 #5447 @kthyng
Congratulations on your new publication @jnnsbrr, and sorry for the long, long delay in this publication! Many thanks to editor @marcosvital and to reviewers @tonyewong, who cruised through a review a long time ago in a timely manner, and @kanishkan91, who heroically jumped in with a review long afterward, for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
